Add Survey, Stratum and PSU to PreySpeciesCategoryCatchData

This function adds biotic PSU definitions (linked to Stratum) and Survey information to PreySpeciesCategoryCatchData.

AddPSUToPreySpeciesCategoryCatch(
  PreySpeciesCategoryCatchData,
  PSUDefinition = c("FunctionParameter", "FunctionInput"),
  PSUDefinitionMethod = c("StationToPSU"),
  BioticPSU = NULL,
  SurveyDefinition = c("FunctionParameter", "FunctionInput"),
  SurveyDefinitionMethod = c("AllStrata", "Table"),
  SurveyTable = data.table::data.table(),
  Survey = NULL,
  StratumPolygon = NULL
)

Arguments

PreySpeciesCategoryCatchData

The PreySpeciesCategoryCatchData data.

PSUDefinition

The method to use for defining the PSUs, one of FunctionParameter to define the PSUs on the fly in this function, or FunctionInput to import PSU process data from a previously run process by BioticPSU.

PSUDefinitionMethod

Character: A string naming the automatic PSU definition method to use. Currently only the option "StationToPSU" is supported, which sets each Station as a PSU. See DefineBioticPSU for details.

BioticPSU

The BioticPSU process data.

SurveyDefinition

The method to use for defining the Survey, one of FunctionParameter to define the Survey on the fly in this function, or FunctionInput to import Survey process data from a previously run process by the input Survey.

SurveyDefinitionMethod

See DefineSurvey

SurveyTable

A table of the two columns Stratum and Survey.

Survey

The Survey process data.

StratumPolygon

The StratumPolygon process data.