Length distribution assigned to each cell of acoustic PSU and Layer

This funciton calculates weighted average of the length distribution of hauls assigned to each acoustic PSU and Layer. The weights are set by BioticAssignmentWeighting.

AssignmentLengthDistribution(LengthDistributionData, BioticAssignment)

Arguments

LengthDistributionData
The LengthDistributionData data.

BioticAssignment
The BioticAssignment process data.

Value

An object of StoX datatype AssignmentLengthDistributionData.

Details

The purpose of the AssignmentLengthDistribution function is to produces one total length distribution for each combination of assigned biotic stations with corresponding weighting variables.

If the biotic station length distributions which shall make up the total length distribution is of type Standard (length distribution as if the complete catch on deck was measured), this will give an implicit weighting by catch as well as towing distance.

If the biotic station length distributions which shall make up the total length distribution is of type Normalized (length distribution as if the complete catch on deck was measured and as if the towing distance had been 1 nautical mile), this will give an implicit weighting by catch.

If the biotic station length distributions which shall make up the total length distribution is of type Percent (sum of percentages for all length groups in the distribution is 100 and the shape of the distribution is the aim), this will give NO implicit weighting.

One total length distribution is calculated as follows:

1) For each biotic station a weighting factor for each station is calculated from the weight variables of the assigned stations:

$$W_s = \frac{w_s}{\sum_{y=1}^{n} w_y}$$

where

\(W_s\) = weighting factor for station \(s\)

\(n\) = number of trawl stations to be combined

\(w_s\) = the value of the weight variable for station \(s\)

\(w_y\) = the value of the weight variable for station \(y\)

2) For each length distribution by biotic station (\(d_s\)), the number or percentage value in each length group is multiplied by \(W_s\):

$$dw_s = W_s \times d_s$$

3) The total length distribution \(d_t\) is finally calculated by adding the numbers in each length interval for all stations where \(n\) is the number of stations to be combined:

$$d_t = \displaystyle\sum_{y=1}^{n} dw_y$$

\(d_t\) is the total length distribution for one assignment.

See also