Weighting of biotic hauls in biotic assignment

This function puts weights to the hauls assigned to acoustic PSUs in BioticAssignment process data.

BioticAssignmentWeighting(
  BioticAssignment,
  WeightingMethod = c("Equal", "NumberOfLengthSamples", "NormalizedTotalWeight",
    "NormalizedTotalNumber", "SumWeightedNumber", "InverseSumWeightedNumber",
    "AcousticDensity"),
  StoxBioticData,
  LengthDistributionData,
  MaxNumberOfLengthSamples = 100,
  NASCData,
  LayerDefinition = c("FunctionParameter", "FunctionInput"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "Table"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  AcousticLayer = NULL,
  AcousticTargetStrength,
  SpeciesLink = data.table::data.table(),
  Radius = double(),
  MinNumberOfEDSUs = integer()
)

Arguments

BioticAssignment

The BioticAssignment process data.

WeightingMethod

Character: A string naming the method to use, one of "Equal", giving weight 1 to all Hauls; "NumberOfLengthSamples", weighting hauls by the number of length samples; "AcousticDensity", weighting by the surrounding NASC converted by the haul length distribution to an acoustic density equivalent; "NormalizedTotalWeight", weighting hauls by the total weight of the catch, normalized by dividing by towed distance; "NormalizedTotalNumber", the same as "NormalizedTotalWeight" but for total number, "SumWeightedNumber", weighting by the summed WeightedNumber of the input LengthDistributionData; and "InverseSumWeightedNumber", weighting by the inverse of the summed WeightedNumber.

StoxBioticData

The StoxBioticData data (defined in RstoxData).

LengthDistributionData

The LengthDistributionData of LengthDistributionType "Standard" or "Normalized", used for WeightingMethod = "SumWeightedNumber" or "InverseSumWeightedNumber", in which case the column WeightedNumber is summed in each Haul (summed over all length groups, which implies that the resolution of length intervals does not matter).

MaxNumberOfLengthSamples

For WeightingMethod = "NumberOfLengthSamples": Values of the number of length samples that exceed MaxNumberOfLengthSamples are set to MaxNumberOfLengthSamples. This avoids giving too high weight to e.g. experimental hauls with particularly large length samples.

NASCData

The NASCData data.

LayerDefinition

The method to use for defining the Layers, one of FunctionParameter to define the Layers on the fly in this function, FunctionInput to import Layer process data from a previously run process using the function DefineAcousticLayer, or PreDefined whihc requires SumNASCData as input.

LayerDefinitionMethod

See DefineAcousticLayer

Resolution

Numeric: A single numeric giving the thickness of the layers.

LayerTable

A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.

AcousticLayer

The AcousticLayer process data.

AcousticTargetStrength

The AcousticTargetStrength process data.

SpeciesLink

A table of the two columns AcousticCategory and SpeciesCategory.

Radius

For WeightingMethod = "AcousticDensity": The radius inside which the average AcousticDensity is calculated.

MinNumberOfEDSUs

For WeightingMethod = "AcousticDensity": The minimum number of EDSUs to link to each Haul, applied if there are less than MinNumberOfEDSUs EDSUs inside the Radius. This can be used to avoid the warning for "no positive NASC inside the specified radius", but only if the NASCData input does not contain EDSUs with no positive NASC for the target species. FilterStoxAcoustic can be used to filter out EDSUs with no positive NASC. It can also happen that the LengthDistributionData contains Hauls with no length distribution for the target species, also resulting in the warning. This can also be solved by filtering. It is in fact recommended to only keep Hauls with the target species in acoustic-trawl models.

Value

An object of StoX data type BioticAssignment.

Details

See also