Assignnment of biotic hauls to acoustic PSUs

This function defines the BioticAssignment process data, linking biotic Hauls with acoustic PSUs.

DefineBioticAssignment(
  processData,
  NASCData,
  StoxBioticData,
  DefinitionMethod = c("Stratum", "Radius", "EllipsoidalDistance"),
  StoxAcousticData,
  AcousticPSU,
  StratumPolygon,
  Radius = double(),
  MinNumStations = integer(),
  RefGCDistance = double(),
  RefTime = "",
  RefBotDepth = double(),
  RefLatitude = double(),
  RefLongitude = double(),
  UseProcessData = FALSE
)

Arguments

processData
The current data produced by a previous instance of the function.

StoxBioticData
A list of StoxBioticData data.

DefinitionMethod
Character: A string naming the method to use, one of "Stratum", which assigns all Hauls to every PSU of each stratum; "Radius", which selects Hauls inside a radius around each PSU????; and "EllipsoidalDistance", which selects Hauls within an ellipsoid defined by axes in space, time, bottom depth *************.

AcousticPSU
A list of AcousticPSU process data.

StratumPolygon
A list of StratumPolygon process data.

UseProcessData
Logical: If TRUE use the existing function output in the process.

AcousticLayer
A list of AcousticLayer process data.

Value

An object of StoX data type BioticAssignment.

Details

This function is awesome and does excellent stuff.

See also

Examples

x <- 1