Biotic PSU

This function defines the BioticPSU process data, linking strata, biotic PSUs and Stations

DefineBioticPSU(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxBioticData,
  DefinitionMethod = c("Manual", "StationToPSU", "DeleteAllPSUs", "ResourceFile"),
  FileName = character()
)

Arguments

processData

The current data produced by a previous instance of the function.

UseProcessData

Logical: If TRUE use the existing function output in the process.

StratumPolygon

The StratumPolygon process data.

StoxBioticData

The StoxBioticData data (defined in RstoxData).

DefinitionMethod

Character: A string naming the method to use, one of "StationToPSU", which sets each Station as a PSU, and "DeleteAllPSUs" to delete all PSUs.

FileName

The path to a resource file from which to read PSUs, in the case that DefinitionMethod is "ResourceFile". Currently, only a project.xml file from StoX 2.7 can be read.

Value

An object of StoX data type BioticPSU.

See also