Define stratum multipolygon

This function reads a GeoJSON file, shapefile or a StoX_multipolygon_WKT file and returns an object of StoX data type StratumPolygon file.

DefineStratumPolygon(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("ResourceFile", "Manual"),
  FileName
)

Arguments

processData
The current data produced by a previous instance of the function.

UseProcessData
Logical: If TRUE use the existing function output in the process.

DefinitionMethod
A string naming the method to use, one of "ResourceFile", to read the file FileName holding the stratum multipolygon, or "Manual" to start off with no strata and create the strata manually in the map of the StoX GUI. Strata can be added, modified and removed in the StoX GUI.

FileName
The path to a GeoJSON file, shapefile (may be a folder) or a StoX_multipolygon_WKT file. Must include file extension.

Value

An object of StoX data type StratumPolygon.

Details

The parameter UseProcessData is always set to TRUE when running a process, and needs to be explicitely set to FALSE to enable reading a file (It's set to FALSE at the moment).

Polygons are expected to have the the following projection: getRstoxBaseDefinitions("proj4string")

References

Additonal information on GeoJSON and shapefiles specification can be found here: https://geojson.org/

https://doc.arcgis.com/en/arcgis-online/reference/shapefiles.htm

See also