Apply the sweep of different gear (and cruise)

This function multiplies the WeightedNumber of a LengthDistributionData by the sweep width given by CompensationTable. The result is a sweep width compensated length distribution (LengthDistributionType starting with "SweepWidthCompensated").

GearDependentCatchCompensation(
  InputDataType = c("LengthDistributionData", "SpeciesCategoryCatchData"),
  LengthDistributionData,
  SpeciesCategoryCatchData,
  CompensationMethod = c("Gear", "Cruise", "GearAndCruise"),
  CompensationTable = data.table::data.table()
)

Arguments

InputDataType

The datatype of the input, one of LengthDistributionData, SpeciesCategoryCatchData.

LengthDistributionData

The LengthDistributionData data.

SpeciesCategoryCatchData

The SpeciesCategoryCatchData data.

CompensationMethod

The method to use for the length dependent catch compensation, i.e. specifying which columns to provide the sweep width for.

CompensationTable

A table of the sweep width per combination of the variables specified in CompensationMethod. Note that all combinations present in the data must be given in the table, as the output should be sweep width compensated for all rows with non-missing WeightedNumber.