Indivduals to distribute abundance to to create super-individuals

This function defines and returns the individuals used in the estimation model, to which to distribute the abundance to create super-individuals.

Individuals(
  StoxBioticData,
  QuantityType = c("Acoustic", "SweptArea"),
  BioticAssignment,
  MeanLengthDistributionData
)

Arguments

StoxBioticData

The StoxBioticData data (defined in RstoxData).

QuantityType

The type of abundance, one of "Acoustic" and "SweptArea".

BioticAssignment

The BioticAssignment process data.

MeanLengthDistributionData

The MeanLengthDistributionData data.

Value

An object of StoX datatype IndividualsData.

Details

The IndividualsData contains variables from StoxBioticData in addition to the columns Stratum and Layer. The Stratum column is not necessarily the actual stratum containing the Haul in which an individual was sampled, but rather the stratum linked to the haul via DefineBioticAssignment for acoustic-trawl models and MeanLengthDistribution for swept-area models. In detail:

DefineAcousticPSU(): Defines acoustic PSUs, which are assigned to a stratum, and that conntainns EDSUs, possibly from other strata than the assigned stratum.

DefineBioticPSU(): Defines biotic PSUs, which are assigned to a stratum, and that conntainns Stations, possibly from other strata than the assigned stratum.

DefineBioticAssignment(): Assignes hauls to acoustic PSUs, possibly from other strata than the stratum assigned to the acoustic PSU. In this datatype it is possible that EDSUs of an acoustic PSU are located in different strata, or even that all the EDSUs of the acoustic PSU are located in another stratum than the assigned stratum; and that the hauls assigned to the acoustic PSU are located in yet another stratum.

In Individuals() the StoxBioticData are merged with BioticAssignment in the case of acoustic-trawl models and with MeanLengthDistributionData in the case of swept-area models, by the Haul identifier stored in the StoxBioticData, the BioticAssignment, and in the Resolution table of the MeanLengthDistributionData. As the hauls may be linked to a different statum than the one containing the haul, as per the description of DefineBioticAssignment() above, the Stratum column of the IndividualsData may not correspond to the actual stratum of the haul.

See also