Mean length distribution data

The MeanLengthDistributionData model data is a list of two elements; the Data, which are similar to LengthDistributionData but added the Layer vertical resolution column and the Stratum and PSU horizontal resolution columns, and summed vertically and averaged horizontally; and the Resolution, which is a table of the five columns Stratum, PSU, Station, Layer and Haul containing the link between the horizontal and vertical resolution variables before summing and averaging.

Details

Data:

Survey	The survey identifier	None	Character	"MainSurvey"	Stratum

The stratum identifier	None	Character	"Stratum1"	PSU	The PSU identifier

None	Character	"PSU1"	Layer	The depth Layer identifier	None

Character	"Layer1"	SpeciesCategory	The species category	None	Character

"Herring"	IndividualTotalLength	Total length (from snout to end of fin), given as the lower end of the interval of width given by LengthResolution	cm	Numeric	14.5

LengthResolution	Resolution of IndividualTotalLength	cm	Numeric	0.5	WeightedNumber

The number of individuals of the length group.	None	Numeric	12	MinLayerDepth	The minimum depth of the depth Layer

m	Numeric	0	MaxLayerDepth	The maximum depth of the depth Layer	m

Numeric	100	MeanLengthDistributionWeight	The weight of the WeightedNumber, equal to the number of Stations of the PSU	None	Numeric

1	LengthDistributionType	The type of length distribution, one of "Normalized", "Standard" and "Percent" (see LengthDistribution)	None	Character	"AreaNumberDensity"

Cruise	Unique Cruise identifier (see StoxAcousticFormat for output from AcousticDensity and StoxBioticFormat for output from SweptAreaDensity)	None	Character	"2021105"	EffectiveTowDistance

Effective tow distance of the Haul	nmi	Numeric	1.5	DateTime	UTC time at start of the EDSU or Station, stored as POSIXct

ISO8601	Character	2020-09-09T01:02:03.456Z	Longitude	Longitude at start of the EDSU or Station	degree east

Numeric	62.5	Latitude	Latitude at start of the EDSU or Station	degree north	Numeric

Resolution:

Variable	Description	Unit	Data type	Example

Stratum	The stratum identifier	None	Character	"Stratum1"

PSU	The PSU identifier	None	Character	"PSU1"

Station	Unique Station identifier	None	Character	"2021105-1"

Layer	The depth Layer identifier	None	Character	"Layer1"

Haul	Unique Haul identifier	None	Character	"2021105-1-2"

Cruise	Unique Cruise identifier (see StoxAcousticFormat for output from AcousticDensity and StoxBioticFormat for output from SweptAreaDensity)	None	Character	"2021105"

EffectiveTowDistance	Effective tow distance of the Haul	nmi	Numeric	1.5

DateTime	UTC time at start of the EDSU or Station, stored as POSIXct	ISO8601	Character	2020-09-09T01:02:03.456Z

Longitude	Longitude at start of the EDSU or Station	degree east	Numeric	62.5

Latitude	Latitude at start of the EDSU or Station	degree north	Numeric	5.1

Gear	Identifier of the gear	None	Character	"3270"

VerticalNetOpening	Vertical span of the net	m	Numeric	23

HorizontalNetOpening	Vertical span of the net	m	Numeric	105

TrawlDoorSpread	Distance between the trawl doors.	m	Numeric	125

See also