Mean species category catch over Stations in each AcousticPSU

This function averages link{SpeciesCategoryCatchData} data horizontally, weighted by the effective towed distance.

MeanSpeciesCategoryCatch(
  SpeciesCategoryCatchData,
  SumSpeciesCategoryCatchData,
  LayerDefinition = c("FunctionParameter", "FunctionInput", "PreDefined"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "Table"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  BioticLayer = NULL,
  SurveyDefinition = c("FunctionParameter", "FunctionInput"),
  SurveyDefinitionMethod = c("AllStrata", "Table"),
  SurveyTable = data.table::data.table(),
  Survey = NULL,
  PSUDefinition = c("FunctionParameter", "FunctionInput"),
  PSUDefinitionMethod = c("StationToPSU", "None"),
  StratumPolygon = NULL,
  BioticPSU = NULL
)

Arguments

SpeciesCategoryCatchData

The SpeciesCategoryCatchData data.

SumSpeciesCategoryCatchData

The SumSpeciesCategoryCatchData data.

LayerDefinition

The method to use for defining the Layers, one of FunctionParameter to define the Layers on the fly in this function, FunctionInput to import Layer process data from a previously run process using the function DefineBioticLayer, or PreDefined whihc requires SumLengthDistributionData as input.

LayerDefinitionMethod

See DefineBioticLayer

Resolution

Numeric: A single numeric giving the thickness of the layers.

LayerTable

A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.

BioticLayer

The BioticLayer process data.

SurveyDefinition

The method to use for defining the Survey, one of FunctionParameter to define the Survey on the fly in this function, or FunctionInput to import Survey process data from a previously run process by the input Survey.

SurveyDefinitionMethod

See DefineSurvey

SurveyTable

A table of the two columns Stratum and Survey.

Survey

The Survey process data.

PSUDefinition

The method to use for defining the PSUs, one of FunctionParameter to define the PSUs on the fly in this function, or FunctionInput to import PSU process data from a previously run process by BioticPSU.

PSUDefinitionMethod

See DefineBioticPSU

StratumPolygon

The StratumPolygon process data.

BioticPSU

The BioticPSU process data.

Value

An MeanSpeciesCategoryCatchData object.