Plot NASCData

Plots a map with cruise line and points with size and (optionally) color representing NASC.

PlotAcousticTrawlSurvey(
  NASCData,
  SumNASCData,
  AcousticPSU,
  LayerDefinition = c("FunctionParameter", "FunctionInput", "PreDefined"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "Table"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  AcousticLayer = NULL,
  ShowOnlyAcousticPSU = FALSE,
  ShowAcousticPSULabel = TRUE,
  UseDefaultAcousticPSULabelSettings = TRUE,
  AcousticPSULabelSize = numeric(),
  AcousticPSULabelColor = character(),
  AcousticPSULabelPosition = c("mean", "atMinLongitude", "atMaxLongitude",
    "atMinLatitude", "atMaxLatitude"),
  AcousticPSULabelHjust = numeric(),
  AcousticPSULabelVjust = numeric(),
  UseDefaultTrackSettings = TRUE,
  TrackColor = character(),
  TrackLineWidth = numeric(),
  TrackPointColor = character(),
  MaxTrackPointSize = numeric(),
  TrackPointShape = numeric(),
  ShowStratumPolygon = FALSE,
  StratumPolygon,
  UseDefaultStratumPolygonSettings = TRUE,
  StratumPolygonColor = character(),
  StratumPolygonBorderColor = character(),
  StratumPolygonBorderLineWidth = 0.5,
  ShowMap = TRUE,
  UseDefaultMapSettings = TRUE,
  LandColor = character(),
  BorderColor = character(),
  OceanColor = character(),
  GridColor = character(),
  UseDefaultAspectSettings = TRUE,
  Zoom = numeric(),
  LongitudeMin = numeric(),
  LongitudeMax = numeric(),
  LatitudeMin = numeric(),
  LatitudeMax = numeric(),
  LongitudeCenter = numeric(),
  LatitudeCenter = numeric(),
  UseDefaultLabelSettings = TRUE,
  Title = character(),
  AxisTitleSize = numeric(),
  AxisTickSize = numeric(),
  LegendTitleSize = numeric(),
  LegendTextSize = numeric(),
  UseDefaultFileSettings = TRUE,
  Format = character(),
  Width = numeric(),
  Height = numeric(),
  DotsPerInch = numeric()
)

Arguments

NASCData

The NASCData data.

SumNASCData

The SumNASCData data.

AcousticPSU

The AcousticPSU process data.

LayerDefinition

The method to use for defining the Layers, one of FunctionParameter to define the Layers on the fly in this function, FunctionInput to import Layer process data generated using the function DefineAcousticLayer, or PreDefined whihc requires SumNASCData as input.

LayerDefinitionMethod

See the argument DefinitionMethod in DefineBioticLayer.

Resolution

Numeric: A single numeric giving the thickness of the layers.

LayerTable

A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.

AcousticLayer

The AcousticLayer process data.

ShowOnlyAcousticPSU

Logical: If TRUE plot only EDSUs tagged to acoustic PSUs.

ShowAcousticPSULabel

Logical: If TRUE show the labels of the acoustic PSUs in the plot.

UseDefaultAcousticPSULabelSettings

Logical: If TRUE (default) use the default settings for acoustic PSU labels in the plots (text position, color and size). Setting this to FALSE will show all acoustic PSU options of the plotting function in the StoX GUI.

AcousticPSULabelSize

Numeric: The size of the AcousticPSU label.

AcousticPSULabelColor

Character: The color of the AcousticPSU label.

AcousticPSULabelPosition

Character: A string determinig where to place the AcousticPSU label, one of "mean", "atMinLongitude", "atMaxLongitude", "atMinLatitude" or "atMaxLatitude".

AcousticPSULabelHjust, AcousticPSULabelVjust

Numeric: The hjust and vjust of the AcousticPSU label for more detailed positioning in addition to the AcousticPSULabelPosition.

UseDefaultTrackSettings

Logical: If TRUE (default) use the default point size settings of the plotting function. Setting this to FALSE will show all point size options of the plotting function in the StoX GUI.

TrackColor

Character: The color to use for the vessel track, defaulted to "black".

TrackLineWidth

Numeric: The width of the track, defaulted to 1.

TrackPointColor

Character: The colors to use when plotting the data points. The colors scale can be given either as vector of colors comprising equally spaced colors of the color scale, or as the name of a color scale function with the first argument being the number of colors. The default is the combined.color. Other options for color scale function are "rainbow", "hcl.colors", "heat.colors", "terrain.colors", "topo.colors" or "cm.colors".

MaxTrackPointSize

Numeric: The maximum size of the points..

TrackPointShape

The shape of the points, as defined for the argument pch in points.

ShowStratumPolygon

Logical: If TRUE include the stratumPolygon. in the plot.

StratumPolygon

The StratumPolygon process data.

UseDefaultStratumPolygonSettings

Logical: If TRUE (default) use the default settings for plotting the StratumPolygon. Setting this to FALSE will show all Stratum plotting options of the plotting function in the StoX GUI.

StratumPolygonColor

Character: The color palette to use for the strata. The default ("hue") is the default HUE color used by ggplot2. See the Palettes section in scale_fill_brewer for a list of options (both Diverging, Qualitative and Sequential color palettes are possible).

StratumPolygonBorderColor

Character: The single color to use for the polygon borders.

StratumPolygonBorderLineWidth

Numeric: The line width of the polygon borders.

ShowMap

Logical: If TRUE include the map in the plot.

UseDefaultMapSettings

Logical: If TRUE (default) use the default settings of the map, specifically LandColor, BorderColor, OceanColor and GridColor. Setting this to FALSE will show all map options of the plotting function in the StoX GUI.

LandColor

Character: The color to use for the land, defaulted to the color used for land in the StoX GUI map (yellow color, rgb(253, 254, 204, maxColorValue = 255)).

BorderColor

Character: The color to use for the borders on land, defaulted to "grey50".

OceanColor

Character: The color to use for the (ocean) background, defaulted to "white", as in the StoX GUI map.

GridColor

Character: The color to use for the longitude/latitude grid lines, defaulted to the color used for the longitude/latitude grid in the StoX GUI map (blue color, rgb(223, 242, 255, maxColorValue = 255).

UseDefaultAspectSettings

Logical: If TRUE (default) use the default aspect settings of the plot, specifically Zoom, LongitudeMin, LongitudeMax, LatitudeMin, LatitudeMax, LongitudeCenter and LatitudeCenter. Setting this to FALSE will show all aspect options of the plotting function in the StoX GUI.

Zoom

Numeric: The zoom of the plot, defaulted to 1, which uses the range of the data as the plot limits. Higher values zooms out and lower values zooms in.

LongitudeMin

Numeric: The minimum longitude limit of the plot, overriding the value derived from Zoom.

LongitudeMax

Numeric: The maximum longitude limit of the plot, overriding the value derived from Zoom.

LatitudeMin

Numeric: The minimum latitude limit of the plot, overriding the value derived from Zoom.

LatitudeMax

Numeric: The maximum latitude limit of the plot, overriding the value derived from Zoom.

LongitudeCenter

Numeric: The longitude of the point around which the plot is zoomed usinng Zoom.

LatitudeCenter

Numeric: The latitude of the point around which the plot is zoomed usinng Zoom.

UseDefaultLabelSettings

Logical: If TRUE (default) use the default text settings of the plot, specifically Title, AxisTitleSize, AxisTickSize, LegendTitleSize and LegendTextSize. Setting this to FALSE will show all text options of the plotting function in the StoX GUI.

Title

Character: The title of the plot, defaulted to no title.

AxisTitleSize

Numeric: The size of the axis titles, defaulted to 20.

AxisTickSize

Numeric: The size of the axis tick marks, defaulted to 20.

LegendTitleSize

Numeric: The size of the legend title, defaulted to 20.

LegendTextSize

Numeric: The size of the legend text, defaulted to 20.

UseDefaultFileSettings

Logical: If TRUE (default) use the default file settings of the output file of the plot, specifically Format, Width, Height and DotsPerInch. Setting this to FALSE will show all file options of the plotting function in the StoX GUI.

Format

Character: The format of the output file from plotting functions, defaulted to "png". Other options are defined by ggsave, including "tiff", "jpeg" and "pdf".

Width

Numeric: The width of the plot in centimeters, defaulted to 17, as per the instructions to authors for submission to the ICES Journal og Marine Science.

Height

Numeric: The height of the plot in centimeters, defaulted to 17.

DotsPerInch

Numeric: The number of dots per inch (DPI), defaulted to 1200.

Value

A PlotAcousticTrawlSurveyData object.