Quantity data

The Quantity model data holds the abundance and biomass of individuals as number and weight, respectively, per Stratum and Layer. The data type contains the tables Data and Resolution with variables given in the following tables:

Details

Data:

Variable	Description	Unit	Data type	Example	Occurrence

Survey	The survey identifier	None	Character	"MainSurvey"	Always

Stratum	The stratum identifier	None	Character	"Stratum1"	Always

Layer	The depth Layer identifier	None	Character	"Layer1"	Always

Beam	Unique Beam identifier	None	Character	"38000/2"	AcousticDensity

SpeciesCategory	The species category	None	Character	"Herring"	Always

IndividualTotalLength	Total length (from snout to end of fin), given as the lower end of the interval of width given by LengthResolution	cm	Numeric	14.5	SweptAreaDensity

LengthResolution	Resolution of IndividualTotalLength	cm	Numeric	0.5	SweptAreaDensity

Frequency	The acoustic frequency of the Beam	hertz	Numeric	38000	AcousticDensity

Abundance	The number of individuals	individuals	Numeric	1234567.8	If MeanDensityType is"AreaNumberDensity" in the input MeanDensityData

Biomass	The total weight of the species	kg	Numeric	123456.7	If DensityType is "AreaWeightDensity" (only relevant when SweptAreaDensityMethod is "TotalCatch" in SweptAreaDensity)

MinLayerDepth	The minimum depth of the depth Layer	m	Numeric	0	Always

MaxLayerDepth	The maximum depth of the depth Layer	m	Numeric	100	Always

Resolution:

Variable	Description	Unit	Data type	Example	Occurrence

Stratum	The stratum identifier	None	Character	"Stratum1"	Always

PSU	The PSU identifier	None	Character	"PSU1"	Always

EDSU	Unique elementary distance sampling unit (EDSU/Log) identifier	None	Character	"2020821/2020-04-24T03:51:26.000Z"	AcousticDensity

Station	Unique Station identifier	None	Character	"2021105-1"	SweptAreaDensity

Layer	The depth Layer identifier	None	Character	"Layer1"	Always

Channel	Unique NASC depth channel identifier	None	Character	"2"	AcousticDensity

Cruise	Unique Cruise identifier (see StoxAcousticFormat for output from AcousticDensity and StoxBioticFormat for output from SweptAreaDensity)	None	Character	"2021105"	Always

EffectiveLogDistance	The effective length of the EDSU	nmi	Numeric	0.09	AcousticDensity

EffectiveTowDistance	Effective tow distance of the Haul	nmi	Numeric	1.5	SweptAreaDensity

DateTime	UTC time at start of the EDSU or Station, stored as POSIXct	ISO8601	Character	2020-09-09T01:02:03.456Z	Always

Longitude	Longitude at start of the EDSU or Station	degree east	Numeric	62.5	Always

Latitude	Latitude at start of the EDSU or Station	degree north	Numeric	5.1	Always

VerticalNetOpening	Vertical span of the net	m	Numeric	23	SweptAreaDensity

HorizontalNetOpening	Vertical span of the net	m	Numeric	105	SweptAreaDensity

TrawlDoorSpread	Distance between the trawl doors.	m	Numeric	125	SweptAreaDensity

See also