Sum NASC

This function summes NASC data vertically.

SumNASC(
  NASCData,
  LayerDefinition = c("FunctionParameter", "FunctionInput"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution",
    "LayerTable"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  AcousticLayer = NULL
)

Arguments

NASCData
The NASCData data.

LayerDefinition
The method to use for defining the Layers, one of FunctionParameter to define the Layers on the fly in this function, or FunctionInput to import Layer process data from a previously run process by AcousticLayer.

LayerDefinitionMethod
See DefineAcousticLayer

Resolution
Numeric: A single numeric giving the thickness of the layers.

LayerTable
A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.

AcousticLayer
The AcousticLayer process data.

Value

A SumNASCData object.

See also