General map plot parameters of RstoxBase

General map plot parameters of RstoxBase

Arguments

LongitudeLimits

The longitude limits of the map plot, defaulted to the range of the longitude of the data.

LatitudeLimits

The latitude limits of the map plot, defaulted to the range of the latitude of the data.

LandColor

The color to use for the land, defaulted to the color used for land in the StoX GUI map (yellow color, rgb(253, 254, 204, maxColorValue = 255)).

BorderColor

The color to use for the borders on land, defaulted to "grey50".

OceanColor

The color to use for the (ocean) background, defaulted to "white", as in the StoX GUI map.

GridColor

The color to use for the longitude/latitude grid lines, defaulted to the color used for the longitude/latitude grid in the StoX GUI map (blue color, rgb(223, 242, 255, maxColorValue = 255).

MaxPointSize

The maximum size of the points, defaulted to 10.

MinPointSize

The minimum size of the points (the size of points of data with value 0), defaulted to 0.5.

TrackColor

The color to use for the vessel track, defaulted to "black".

TrackSize

The size of the track points, defaulted to 1.