General ploting parameters of RstoxBase

General ploting parameters of RstoxBase

Arguments

UseDefaultColorSettings

Logical: If TRUE (default) use the default color settings of the plotting function. Setting this to FALSE will show all color options of the plotting function in the StoX GUI.

UseDefaultSizeSettings

Logical: If TRUE (default) use the default point size settings of the plotting function. Setting this to FALSE will show all point size options of the plotting function in the StoX GUI.

UseDefaultAspectSettings

Logical: If TRUE (default) use the default aspect settings of the plotting function (width, height, etc.). Setting this to FALSE will show all aspect options of the plotting function in the StoX GUI.

UseDefaultTextSettings

Logical: If TRUE (default) use the default text settings of the plotting function. Setting this to FALSE will show all text options of the plotting function in the StoX GUI.

UseDefaultFileSettings

Logical: If TRUE (default) use the default file settings of the plotting function. Setting this to FALSE will show all file options of the plotting function in the StoX GUI.

Zoom

The zoom of the plot, defaulted to 1, which uses the range of the data as the plot limits. Higher values zooms out and lower values zooms in.

Title

The title of the plot, defaulted to no title.

AxisTitleSize

The size of the axis titles, defaulted to 20.

AxisTickSize

The size of the axis tick marks, defaulted to 20.

LegendTitleSize

The size of the legend title, defaulted to 20.

LegendTextSize

The size of the legend text, defaulted to 20.

Format

The format of the output file from plotting functions, defaulted to "png". Other options are defined by ggsave, including "tiff", "jpeg" and "pdf".

Width

The width of the plot in centimeters, defaulted to 17, as per the ICES Journal og Marine Science author recommendations.

Height

The height of the plot in centimeters, defaulted to 17.

DotsPerInch

The number of dots per inch (DPI), defaulted to 1200.