Get RstoxBase definitions

This function gets vital definitions from the RstoxBase environment.

getRstoxBaseDefinitions(name = NULL, ...)

Arguments

name

An optional string vector denoting which definitions to extract.

...

values overriding the values of definitions.

Value

A list of definitions.

Examples

getRstoxBaseDefinitions()
#> $allResolution
#> $allResolution$NASCData
#> [1] "NASC"
#> 
#> $allResolution$SumNASCData
#> [1] "NASC"
#> 
#> $allResolution$MeanNASCData
#> [1] "NASC"
#> 
#> $allResolution$LengthDistributionData
#> [1] "LengthDistribution"
#> 
#> $allResolution$SumLengthDistributionData
#> [1] "LengthDistribution"
#> 
#> $allResolution$MeanLengthDistributionData
#> [1] "LengthDistribution"
#> 
#> $allResolution$DensityData
#> [1] "Density"
#> 
#> $allResolution$MeanDensityData
#> [1] "Density"
#> 
#> $allResolution$QuantityData
#> [1] "Quantity"
#> 
#> 
#> $dataTypeDefinition
#> $dataTypeDefinition$NASCData
#> $dataTypeDefinition$NASCData$horizontalResolution
#> [1] "EDSU"
#> 
#> $dataTypeDefinition$NASCData$verticalResolution
#> [1] "Channel"
#> 
#> $dataTypeDefinition$NASCData$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$NASCData$categoryVariable
#> [1] "AcousticCategory"
#> 
#> $dataTypeDefinition$NASCData$groupingVariables
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$NASCData$data
#> [1] "NASC"
#> 
#> $dataTypeDefinition$NASCData$verticalRawDimension
#> [1] "MinChannelDepth" "MaxChannelDepth"
#> 
#> $dataTypeDefinition$NASCData$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$NASCData$weighting
#> [1] "NASCWeight"
#> 
#> $dataTypeDefinition$NASCData$type
#> [1] "ChannelReferenceType"
#> 
#> $dataTypeDefinition$NASCData$other
#> [1] "ChannelReferenceDepth" "ChannelReferenceTilt"  "Cruise"               
#> [4] "EffectiveLogDistance"  "DateTime"              "Longitude"            
#> [7] "Latitude"             
#> 
#> 
#> $dataTypeDefinition$SumNASCData
#> $dataTypeDefinition$SumNASCData$Data
#> $dataTypeDefinition$SumNASCData$Data$horizontalResolution
#> [1] "EDSU"
#> 
#> $dataTypeDefinition$SumNASCData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$SumNASCData$Data$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$SumNASCData$Data$categoryVariable
#> [1] "AcousticCategory"
#> 
#> $dataTypeDefinition$SumNASCData$Data$groupingVariables
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$SumNASCData$Data$data
#> [1] "NASC"
#> 
#> $dataTypeDefinition$SumNASCData$Data$verticalRawDimension
#> NULL
#> 
#> $dataTypeDefinition$SumNASCData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$SumNASCData$Data$weighting
#> [1] "SumNASCWeight"
#> 
#> $dataTypeDefinition$SumNASCData$Data$type
#> [1] "ChannelReferenceType"
#> 
#> $dataTypeDefinition$SumNASCData$Data$other
#> [1] "ChannelReferenceDepth" "ChannelReferenceTilt"  "Cruise"               
#> [4] "EffectiveLogDistance"  "DateTime"              "Longitude"            
#> [7] "Latitude"             
#> 
#> 
#> $dataTypeDefinition$SumNASCData$Resolution
#> $dataTypeDefinition$SumNASCData$Resolution$horizontalResolution
#> [1] "EDSU"
#> 
#> $dataTypeDefinition$SumNASCData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> 
#> 
#> $dataTypeDefinition$MeanNASCData
#> $dataTypeDefinition$MeanNASCData$Data
#> $dataTypeDefinition$MeanNASCData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$MeanNASCData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$categoryVariable
#> [1] "AcousticCategory"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$groupingVariables
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$data
#> [1] "NASC"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$verticalRawDimension
#> NULL
#> 
#> $dataTypeDefinition$MeanNASCData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$weighting
#> [1] "MeanNASCWeight"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$type
#> [1] "ChannelReferenceType"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$other
#> [1] "ChannelReferenceDepth" "ChannelReferenceTilt" 
#> 
#> 
#> $dataTypeDefinition$MeanNASCData$Resolution
#> $dataTypeDefinition$MeanNASCData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "EDSU"   
#> 
#> $dataTypeDefinition$MeanNASCData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> $dataTypeDefinition$MeanNASCData$Resolution$other
#> [1] "Cruise"               "EffectiveLogDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"            
#> 
#> 
#> 
#> $dataTypeDefinition$LengthDistributionData
#> $dataTypeDefinition$LengthDistributionData$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$LengthDistributionData$verticalResolution
#> [1] "Haul"
#> 
#> $dataTypeDefinition$LengthDistributionData$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$LengthDistributionData$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$LengthDistributionData$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$LengthDistributionData$data
#> [1] "WeightedNumber"
#> 
#> $dataTypeDefinition$LengthDistributionData$verticalRawDimension
#> [1] "MinHaulDepth" "MaxHaulDepth"
#> 
#> $dataTypeDefinition$LengthDistributionData$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$LengthDistributionData$weighting
#> [1] "LengthDistributionWeight"
#> 
#> $dataTypeDefinition$LengthDistributionData$type
#> [1] "LengthDistributionType"
#> 
#> $dataTypeDefinition$LengthDistributionData$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"             "Gear"                
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> $dataTypeDefinition$SumLengthDistributionData
#> $dataTypeDefinition$SumLengthDistributionData$Data
#> $dataTypeDefinition$SumLengthDistributionData$Data$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$data
#> [1] "WeightedNumber"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$verticalRawDimension
#> NULL
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$weighting
#> [1] "SumLengthDistributionWeight"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$type
#> [1] "LengthDistributionType"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"            
#> 
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Resolution
#> $dataTypeDefinition$SumLengthDistributionData$Resolution$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Resolution$verticalResolution
#> [1] "Layer" "Haul" 
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Resolution$other
#> [1] "Gear"                 "VerticalNetOpening"   "HorizontalNetOpening"
#> [4] "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$MeanLengthDistributionData
#> $dataTypeDefinition$MeanLengthDistributionData$Data
#> $dataTypeDefinition$MeanLengthDistributionData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$data
#> [1] "WeightedNumber"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$verticalRawDimension
#> NULL
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$weighting
#> [1] "MeanLengthDistributionWeight"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$type
#> [1] "LengthDistributionType"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Resolution
#> $dataTypeDefinition$MeanLengthDistributionData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "Station"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Resolution$verticalResolution
#> [1] "Layer" "Haul" 
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Resolution$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"             "Gear"                
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData
#> $dataTypeDefinition$SpeciesCategoryCatchData$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$verticalResolution
#> [1] "Haul"
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$groupingVariables
#> NULL
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$data
#>             Weight             Number 
#> "TotalCatchWeight" "TotalCatchNumber" 
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$verticalRawDimension
#> [1] "MinHaulDepth" "MaxHaulDepth"
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$weighting
#> [1] "SpeciesCategoryCatchWeight"
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$type
#> [1] "SpeciesCategoryCatchType"
#> 
#> $dataTypeDefinition$SpeciesCategoryCatchData$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"             "Gear"                
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$groupingVariables
#> NULL
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$data
#>             Weight             Number 
#> "TotalCatchWeight" "TotalCatchNumber" 
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$verticalRawDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$weighting
#> [1] "SumSpeciesCategoryCatchWeight"
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$type
#> [1] "SpeciesCategoryCatchType"
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Data$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"            
#> 
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Resolution
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Resolution$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Resolution$verticalResolution
#> [1] "Layer" "Haul" 
#> 
#> $dataTypeDefinition$SumSpeciesCategoryCatchData$Resolution$other
#> [1] "Gear"                 "VerticalNetOpening"   "HorizontalNetOpening"
#> [4] "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$groupingVariables
#> NULL
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$data
#>             Weight             Number 
#> "TotalCatchWeight" "TotalCatchNumber" 
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$verticalRawDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$weighting
#> [1] "MeanSpeciesCategoryCatchWeight"
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$type
#> [1] "SpeciesCategoryCatchType"
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Resolution
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "Station"
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Resolution$verticalResolution
#> [1] "Layer" "Haul" 
#> 
#> $dataTypeDefinition$MeanSpeciesCategoryCatchData$Resolution$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"             "Gear"                
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData
#> $dataTypeDefinition$AssignmentLengthDistributionData$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$data
#> [1] "WeightedNumber"
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$weighting
#> NULL
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$type
#> [1] "LengthDistributionType"
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$other
#> [1] "NumberOfAssignedHauls"          "AllHaulsHaveAllSpeciesCategory"
#> [3] "AllHaulsHaveAnySpeciesCategory"
#> 
#> 
#> $dataTypeDefinition$DensityData
#> $dataTypeDefinition$DensityData$Data
#> $dataTypeDefinition$DensityData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$DensityData$Data$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$DensityData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$DensityData$Data$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$DensityData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$DensityData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"      "Beam"                 
#> [4] "Frequency"            
#> 
#> $dataTypeDefinition$DensityData$Data$groupingVariables_biotic
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$DensityData$Data$groupingVariables_acoustic
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$DensityData$Data$data
#> [1] "Density"
#> 
#> $dataTypeDefinition$DensityData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$DensityData$Data$weighting
#> [1] "DensityWeight"
#> 
#> $dataTypeDefinition$DensityData$Data$type
#> [1] "DensityType"
#> 
#> $dataTypeDefinition$DensityData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$DensityData$Resolution
#> $dataTypeDefinition$DensityData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "EDSU"    "Station"
#> 
#> $dataTypeDefinition$DensityData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> $dataTypeDefinition$DensityData$Resolution$other
#> [1] "Cruise"               "EffectiveLogDistance" "EffectiveTowDistance"
#> [4] "DateTime"             "Longitude"            "Latitude"            
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$MeanDensityData
#> $dataTypeDefinition$MeanDensityData$Data
#> $dataTypeDefinition$MeanDensityData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$horizontalResolution
#> [1] "Stratum"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"      "Beam"                 
#> [4] "Frequency"            
#> 
#> $dataTypeDefinition$MeanDensityData$Data$groupingVariables_biotic
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$MeanDensityData$Data$groupingVariables_acoustic
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$data
#> [1] "Density"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$weighting
#> [1] "MeanDensityWeight"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$type
#> [1] "DensityType"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$MeanDensityData$Resolution
#> $dataTypeDefinition$MeanDensityData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "EDSU"    "Station"
#> 
#> $dataTypeDefinition$MeanDensityData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> $dataTypeDefinition$MeanDensityData$Resolution$other
#> [1] "Cruise"               "EffectiveLogDistance" "EffectiveTowDistance"
#> [4] "DateTime"             "Longitude"            "Latitude"            
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$QuantityData
#> $dataTypeDefinition$QuantityData$Data
#> $dataTypeDefinition$QuantityData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$QuantityData$Data$horizontalResolution
#> [1] "Stratum"
#> 
#> $dataTypeDefinition$QuantityData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$QuantityData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$QuantityData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"      "Beam"                 
#> [4] "Frequency"            
#> 
#> $dataTypeDefinition$QuantityData$Data$groupingVariables_biotic
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$QuantityData$Data$groupingVariables_acoustic
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$QuantityData$Data$data
#> [1] "Abundance" "Biomass"  
#> 
#> $dataTypeDefinition$QuantityData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$QuantityData$Data$weighting
#> NULL
#> 
#> $dataTypeDefinition$QuantityData$Data$type
#> NULL
#> 
#> $dataTypeDefinition$QuantityData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$QuantityData$Resolution
#> $dataTypeDefinition$QuantityData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "EDSU"    "Station"
#> 
#> $dataTypeDefinition$QuantityData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> $dataTypeDefinition$QuantityData$Resolution$other
#> [1] "Cruise"               "EffectiveLogDistance" "EffectiveTowDistance"
#> [4] "DateTime"             "Longitude"            "Latitude"            
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$IndividualsData
#> $dataTypeDefinition$IndividualsData$Data
#> $dataTypeDefinition$IndividualsData$Data$horizontalResolution
#> [1] "Stratum"
#> 
#> $dataTypeDefinition$IndividualsData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$IndividualsData$Data$data
#> NULL
#> 
#> $dataTypeDefinition$IndividualsData$Data$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$IndividualsData$Data$weighting
#> NULL
#> 
#> $dataTypeDefinition$IndividualsData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$IndividualsData$VariableNames
#> list()
#> 
#> 
#> $dataTypeDefinition$SuperIndividualsData
#> $dataTypeDefinition$SuperIndividualsData$Data
#> $dataTypeDefinition$SuperIndividualsData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$horizontalResolution
#> [1] "Stratum"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$groupingVariables
#>  [1] "CruiseKey"             "StationKey"            "HaulKey"              
#>  [4] "SpeciesCategoryKey"    "SampleKey"             "IndividualKey"        
#>  [7] "IndividualTotalLength" "LengthResolution"      "IndividualRoundWeight"
#> [10] "Beam"                  "Frequency"            
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$data
#> [1] "Abundance" "Biomass"  
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$weighting
#> NULL
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$SuperIndividualsData$VariableNames
#> list()
#> 
#> 
#> $dataTypeDefinition$BioticAssignment
#> $dataTypeDefinition$BioticAssignment$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$BioticAssignment$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$BioticAssignment$categoryVariable
#> [1] "Haul"
#> 
#> $dataTypeDefinition$BioticAssignment$weighting
#> [1] "WeightingFactor"
#> 
#> $dataTypeDefinition$BioticAssignment$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$BioticPSU
#> $dataTypeDefinition$BioticPSU$Stratum_PSU
#> $dataTypeDefinition$BioticPSU$Stratum_PSU$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> 
#> $dataTypeDefinition$BioticPSU$Station_PSU
#> $dataTypeDefinition$BioticPSU$Station_PSU$horizontalResolution
#> [1] "Station" "PSU"    
#> 
#> 
#> 
#> $dataTypeDefinition$AcousticPSU
#> $dataTypeDefinition$AcousticPSU$Stratum_PSU
#> $dataTypeDefinition$AcousticPSU$Stratum_PSU$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> 
#> $dataTypeDefinition$AcousticPSU$EDSU_PSU
#> $dataTypeDefinition$AcousticPSU$EDSU_PSU$horizontalResolution
#> [1] "EDSU" "PSU" 
#> 
#> 
#> $dataTypeDefinition$AcousticPSU$PSUByTime
#> $dataTypeDefinition$AcousticPSU$PSUByTime$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$AcousticPSU$PSUByTime$categoryVariable
#> [1] "Cruise"
#> 
#> $dataTypeDefinition$AcousticPSU$PSUByTime$groupingVariables
#> [1] "StartDateTime" "StopDateTime" 
#> 
#> 
#> 
#> 
#> $dataTypeUnits
#> $dataTypeUnits$NASCData
#> $dataTypeUnits$NASCData$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$NASCData$NASC
#> [1] "sA"
#> 
#> $dataTypeUnits$NASCData$MinChannelDepth
#> [1] "m"
#> 
#> $dataTypeUnits$NASCData$MaxChannelDepth
#> [1] "m"
#> 
#> $dataTypeUnits$NASCData$ChannelReferenceDepth
#> [1] "m"
#> 
#> $dataTypeUnits$NASCData$ChannelReferenceTilt
#> [1] "degrees"
#> 
#> $dataTypeUnits$NASCData$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$NASCData$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$NASCData$Latitude
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$SumNASCData
#> $dataTypeUnits$SumNASCData$Data
#> $dataTypeUnits$SumNASCData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$SumNASCData$Data$NASC
#> [1] "sA"
#> 
#> $dataTypeUnits$SumNASCData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumNASCData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumNASCData$Data$ChannelReferenceDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumNASCData$Data$ChannelReferenceTilt
#> [1] "degrees"
#> 
#> $dataTypeUnits$SumNASCData$Data$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$SumNASCData$Data$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$SumNASCData$Data$Latitude
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$SumNASCData$Resolution
#> list()
#> 
#> 
#> $dataTypeUnits$MeanNASCData
#> $dataTypeUnits$MeanNASCData$Data
#> $dataTypeUnits$MeanNASCData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$MeanNASCData$Data$NASC
#> [1] "sA"
#> 
#> $dataTypeUnits$MeanNASCData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanNASCData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanNASCData$Data$ChannelReferenceDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanNASCData$Data$ChannelReferenceTilt
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$MeanNASCData$Resolution
#> $dataTypeUnits$MeanNASCData$Resolution$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanNASCData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanNASCData$Resolution$Latitude
#> [1] "degrees"
#> 
#> 
#> 
#> $dataTypeUnits$LengthDistributionData
#> $dataTypeUnits$LengthDistributionData$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$LengthDistributionData$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$LengthDistributionData$WeightedNumber
#> [1] "individual"
#> 
#> $dataTypeUnits$LengthDistributionData$MinHaulDepth
#> [1] "m"
#> 
#> $dataTypeUnits$LengthDistributionData$MaxHaulDepth
#> [1] "m"
#> 
#> $dataTypeUnits$LengthDistributionData$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$LengthDistributionData$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$LengthDistributionData$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$LengthDistributionData$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$LengthDistributionData$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$LengthDistributionData$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$SumLengthDistributionData
#> $dataTypeUnits$SumLengthDistributionData$Data
#> $dataTypeUnits$SumLengthDistributionData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$WeightedNumber
#> [1] "individual"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$Latitude
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$SumLengthDistributionData$Resolution
#> $dataTypeUnits$SumLengthDistributionData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$MeanLengthDistributionData
#> $dataTypeUnits$MeanLengthDistributionData$Data
#> $dataTypeUnits$MeanLengthDistributionData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$WeightedNumber
#> [1] "individual"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$Latitude
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$AssignmentLengthDistributionData
#> $dataTypeUnits$AssignmentLengthDistributionData$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$AssignmentLengthDistributionData$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$AssignmentLengthDistributionData$WeightedNumber
#> [1] "individual"
#> 
#> 
#> $dataTypeUnits$DensityData
#> $dataTypeUnits$DensityData$Data
#> $dataTypeUnits$DensityData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$DensityData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$DensityData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$DensityData$Data$Density
#> [1] "nautical_areal_number_density"
#> 
#> $dataTypeUnits$DensityData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$DensityData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$DensityData$Resolution
#> $dataTypeUnits$DensityData$Resolution$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$DensityData$Resolution$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$DensityData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$DensityData$Resolution$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$DensityData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$DensityData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$DensityData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$MeanDensityData
#> $dataTypeUnits$MeanDensityData$Data
#> $dataTypeUnits$MeanDensityData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$MeanDensityData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$MeanDensityData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$MeanDensityData$Data$Density
#> [1] "nautical_areal_number_density"
#> 
#> $dataTypeUnits$MeanDensityData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanDensityData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$MeanDensityData$Resolution
#> $dataTypeUnits$MeanDensityData$Resolution$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$QuantityData
#> $dataTypeUnits$QuantityData$Data
#> $dataTypeUnits$QuantityData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$QuantityData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$QuantityData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$QuantityData$Data$Abundance
#> [1] "ind"
#> 
#> $dataTypeUnits$QuantityData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$QuantityData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$QuantityData$Resolution
#> $dataTypeUnits$QuantityData$Resolution$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$QuantityData$Resolution$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$QuantityData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$QuantityData$Resolution$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$QuantityData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$QuantityData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$QuantityData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$IndividualsData
#> list()
#> 
#> $dataTypeUnits$SuperIndividualsData
#> $dataTypeUnits$SuperIndividualsData$Data
#> $dataTypeUnits$SuperIndividualsData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$IndividualRoundWeight
#> [1] "g"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$Abundance
#> [1] "ind"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$Biomass
#> [1] "g"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$SuperIndividualsData$VariableNames
#> list()
#> 
#> 
#> $dataTypeUnits$BioticAssignment
#> list()
#> 
#> $dataTypeUnits$BioticPSU
#> list()
#> 
#> $dataTypeUnits$AcousticPSU
#> list()
#> 
#> 
#> $defaultEstimationMethod
#> $defaultEstimationMethod$Regression
#> $defaultEstimationMethod$Regression$SimpleLinear
#> [1] "Linear"
#> 
#> $defaultEstimationMethod$Regression$Power
#> [1] "LogLogLinear"
#> 
#> 
#> 
#> $emptyStratumPolygon
#> An object of class "SpatialPolygonsDataFrame"
#> Slot "data":
#> data frame with 0 columns and 0 rows
#> 
#> Slot "polygons":
#> list()
#> 
#> Slot "plotOrder":
#> integer(0)
#> 
#> Slot "bbox":
#>              min           max
#> x -1.797693e+308 1.797693e+308
#> y -1.797693e+308 1.797693e+308
#> 
#> Slot "proj4string":
#> Coordinate Reference System:
#> Deprecated Proj.4 representation: NA 
#> Warning: CRS object has no comment
#> 
#> 
#> $estimationFunctions
#> $estimationFunctions$Regression
#> $estimationFunctions$Regression$SimpleLinear
#> function (dependentVariable, independentVariable, data) 
#> {
#>     fit <- stats::lm(get(dependentVariable) ~ get(independentVariable), 
#>         data = data, na.action = stats::na.exclude)
#>     return(fit)
#> }
#> <bytecode: 0x7fa87cc010e8>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $estimationFunctions$Regression$Power
#> function (dependentVariable, independentVariable, data) 
#> {
#>     is0 <- which(data[, c(dependentVariable, independentVariable), 
#>         with = FALSE] == 0, arr.ind = TRUE)
#>     if (NROW(is0)) {
#>         stop("There are zeros in the data, causing -Inf in the regression. This can happen when regressing IndividualTotalLength in SuperIndividualsData, where the length resolution may have been reduced to that some individuals end up with IndividualTotalLength = 0.")
#>     }
#>     fit <- stats::lm(log(get(dependentVariable)) ~ log(get(independentVariable)), 
#>         data = data, na.action = stats::na.exclude)
#>     fit$coefficients[1] <- exp(fit$coefficients[1])
#>     return(fit)
#> }
#> <bytecode: 0x7fa87cc04770>
#> <environment: 0x7fa87cc07eb0>
#> 
#> 
#> 
#> $getPSUPrefix
#> function (PSUType) 
#> {
#>     if (PSUType == "Acoustic") {
#>         prefix <- "PSU"
#>     }
#>     else if (PSUType == "Biotic") {
#>         prefix <- "PSU"
#>     }
#>     else {
#>         stop("PSUType must be one of \"Acoustic\" or \"Biotic\"")
#>     }
#>     return(prefix)
#> }
#> <bytecode: 0x7fa87cc06268>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $getSSULabel
#> function (PSUType) 
#> {
#>     if (PSUType == "Acoustic") {
#>         SSULabel <- "EDSU"
#>     }
#>     else if (PSUType == "Biotic") {
#>         SSULabel <- "Station"
#>     }
#>     else {
#>         stop("PSUType must be one of \"Acoustic\" or \"Biotic\"")
#>     }
#>     return(SSULabel)
#> }
#> <bytecode: 0x7fa87cc05a50>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $getStationLevel
#> function (PSUType) 
#> {
#>     if (PSUType == "Acoustic") {
#>         SSULabel <- "Log"
#>     }
#>     else if (PSUType == "Biotic") {
#>         SSULabel <- "Station"
#>     }
#>     else {
#>         stop("PSUType must be one of \"Acoustic\" or \"Biotic\"")
#>     }
#>     return(SSULabel)
#> }
#> <bytecode: 0x7fa87cc05158>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $implementedModelClasses
#> [1] "AcousticTargetStrength" "Regression"            
#> 
#> $modelFunctions
#> $modelFunctions$AcousticTargetStrength
#> $modelFunctions$AcousticTargetStrength$LengthDependent
#> function (midIndividualTotalLength, TargetStrength0, LengthExponent, 
#>     Depth) 
#> {
#>     TargetStrength0 + LengthExponent * log10(midIndividualTotalLength)
#> }
#> <bytecode: 0x7fa87cc02888>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $modelFunctions$AcousticTargetStrength$LengthAndDepthDependent
#> function (midIndividualTotalLength, TargetStrength0, LengthExponent, 
#>     DepthExponent, Depth) 
#> {
#>     TargetStrength0 + LengthExponent * log10(midIndividualTotalLength) + 
#>         DepthExponent * log10(1 + Depth/10)
#> }
#> <bytecode: 0x7fa87cc023b8>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $modelFunctions$AcousticTargetStrength$LengthExponent
#> function (midIndividualTotalLength, LengthExponent) 
#> {
#>     LengthExponent * log10(midIndividualTotalLength)
#> }
#> <bytecode: 0x7fa87cc01d28>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $modelFunctions$AcousticTargetStrength$TargetStrengthByLength
#> [1] NA
#> 
#> 
#> $modelFunctions$Regression
#> $modelFunctions$Regression$SimpleLinear
#> function (independentVariable, parameters) 
#> {
#>     parameters$Intercept + parameters$Slope * get(independentVariable)
#> }
#> <bytecode: 0x7fa87cc01a50>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $modelFunctions$Regression$Power
#> function (independentVariable, parameters) 
#> {
#>     parameters$Factor * independentVariable^parameters$Exponent
#> }
#> <bytecode: 0x7fa87cc01548>
#> <environment: 0x7fa87cc07eb0>
#> 
#> 
#> 
#> $modelNames
#> $modelNames$AcousticTargetStrength
#> [1] "LengthDependent"         "LengthAndDepthDependent"
#> [3] "LengthExponent"          "TargetStrengthByLength" 
#> 
#> $modelNames$Regression
#> $modelNames$Regression[[1]]
#> [1] "SimpleLinear"
#> 
#> $modelNames$Regression[[2]]
#> [1] "Power"
#> 
#> 
#> 
#> $modelParameters
#> $modelParameters$AcousticTargetStrength
#> $modelParameters$AcousticTargetStrength$LengthDependent
#> [1] "TargetStrength0" "LengthExponent" 
#> 
#> $modelParameters$AcousticTargetStrength$LengthAndDepthDependent
#> [1] "TargetStrength0" "LengthExponent"  "DepthExponent"  
#> 
#> $modelParameters$AcousticTargetStrength$LengthExponent
#> [1] "LengthExponent"
#> 
#> $modelParameters$AcousticTargetStrength$TargetStrengthByLength
#> [1] "TargetStrength" "TotalLength"   
#> 
#> 
#> $modelParameters$Regression
#> $modelParameters$Regression$SimpleLinear
#> [1] "Intercept" "Slope"    
#> 
#> $modelParameters$Regression$Power
#> [1] "Factor"   "Exponent"
#> 
#> 
#> 
#> $modelTypes
#> $modelTypes$AcousticTargetStrength
#> $modelTypes$AcousticTargetStrength$LengthDependent
#> [1] "Function"
#> 
#> $modelTypes$AcousticTargetStrength$LengthAndDepthDependent
#> [1] "Function"
#> 
#> $modelTypes$AcousticTargetStrength$LengthExponent
#> [1] "Function"
#> 
#> $modelTypes$AcousticTargetStrength$TargetStrengthByLength
#> [1] "Table"
#> 
#> 
#> $modelTypes$Regression
#> $modelTypes$Regression$SimpleLinear
#> [1] "Function"
#> 
#> $modelTypes$Regression$Power
#> [1] "Function"
#> 
#> 
#> 
#> $nauticalMileInMeters
#> [1] 1852
#> 
#> $proj4string_laea
#> [1] "+proj=laea +no_defs +ellps=WGS84 +towgs84=0,0,0"
#> 
#> $proj4string_longlat
#> [1] "+proj=longlat +no_defs +ellps=WGS84 +towgs84=0,0,0"
#> 
#> $PSUByTime
#> $PSUByTime$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $PSUByTime$categoryVariable
#> [1] "Cruise"
#> 
#> $PSUByTime$groupingVariables
#> [1] "StartDateTime" "StopDateTime" 
#> 
#> 
#> $RemoveMissingValuesWarning
#> function (TargetVariable) 
#> {
#>     paste0("StoX: The TargetVariable (", TargetVariable, ") has missing values. Use RemoveMissingValues = TRUE with extreme caution!!! The Baseline function ImputeSuperIndividuals can be used to fill in missing information from other super-individuals.")
#> }
#> <bytecode: 0x7fa87cc08620>
#> <environment: 0x7fa87cc07eb0>
#> 
#> $reportFunctions
#>      functionName packageName weighted weightingParameter multiple
#>  1:   summaryStox   RstoxBase    FALSE                        TRUE
#>  2:           sum        base    FALSE                       FALSE
#>  3:          mean        base    FALSE                       FALSE
#>  4: weighted.mean       stats     TRUE                  w    FALSE
#>  5:        median       stats    FALSE                       FALSE
#>  6:           min        base    FALSE                       FALSE
#>  7:           max        base    FALSE                       FALSE
#>  8:            sd       stats    FALSE                       FALSE
#>  9:           var       stats    FALSE                       FALSE
#> 10:            cv   RstoxBase    FALSE                       FALSE
#> 
#> $resolutionClasses
#> $resolutionClasses$NASC
#> $resolutionClasses$NASC$vertical
#> [1] "Layer"   "Channel"
#> 
#> $resolutionClasses$NASC$horizontal
#> [1] "Stratum" "PSU"     "EDSU"   
#> 
#> 
#> $resolutionClasses$LengthDistribution
#> $resolutionClasses$LengthDistribution$vertical
#> [1] "Layer" "Haul" 
#> 
#> $resolutionClasses$LengthDistribution$horizontal
#> [1] "Stratum" "PSU"     "Station"
#> 
#> 
#> $resolutionClasses$Density
#> $resolutionClasses$Density$vertical
#> [1] "Layer"
#> 
#> $resolutionClasses$Density$horizontal
#> [1] "Stratum" "PSU"    
#> 
#> 
#> $resolutionClasses$Quantity
#> $resolutionClasses$Quantity$vertical
#> [1] "Layer"
#> 
#> $resolutionClasses$Quantity$horizontal
#> [1] "Stratum"
#> 
#> 
#> 
#> $seedSequenceLength
#> [1] 1e+07
#>