Modify a parameter of a process of a list representation of a project.xml file from StoX <= 2.7.

Modify a parameter of a process of a list representation of a project.xml file from StoX <= 2.7.

modifyParameter_ProjectXMLList(
  projectXMLList,
  modelName,
  processName,
  parameterName,
  newParameterValue,
  functionName = NULL
)

Arguments

projectXMLList

A list as read by readProjectXMLToList.

modelName

The name of the model (possible values are "baseline", "baseline-report", "r" and "r-report").

processName

he name of the process.

parameterName

The name of the parameter.

newParameterValue

The new value of the parameter.

functionName

The name of the function of the process to add.