Weighting of biotic hauls in biotic assignment

This function puts weights to the hauls assigned to acoustic PSUs in BioticAssignment process data.

BioticAssignmentWeighting(
  BioticAssignment,
  WeightingMethod = c("Equal", "NumberOfLengthSamples", "NormalizedTotalWeight",
    "NormalizedTotalCount", "SumWeightedCount", "InverseSumWeightedCount"),
  StoxBioticData,
  LengthDistributionData,
  MaxNumberOfLengthSamples = 100,
  StoxAcousticData,
  Radius = double(),
  LengthExponent = double()
)

Arguments

BioticAssignment
The BioticAssignment process data.

WeightingMethod
Character: A string naming the method to use, one of "Equal", giving weight 1 to all Hauls; "NumberOfLengthSamples", weighting hauls by the number of length samples; "NASC", weighting by the surrounding NASC converted by the haul length distribution to a density equivalent; "NormalizedTotalWeight", weighting hauls by the total weight of the catch, normalized by dividing by towed distance; "NormalizedTotalCount", the same as "NormalizedTotalWeight" but for total count, "SumWeightedCount", weighting by the summed WeightedCount of the input LengthDistributionData; and "InverseSumWeightedCount", weighting by the inverse of the summed WeightedCount.

StoxBioticData
The StoxBioticData data (defined in RstoxData).

LengthDistributionData
The LengthDistributionData data.

MaxNumberOfLengthSamples
For WeightingMethod = "NumberOfLengthSamples": Values of the number of length samples that exceed MaxNumberOfLengthSamples are set to MaxNumberOfLengthSamples. This avoids giving too high weight to e.g. experimental hauls with particularly large length samples.

StoxAcousticData
The StoxAcousticData data (defined in RstoxData).

Radius
For WeightingMethod = "NASC": The radius inside which the average NASC is calculated.

LengthExponent
For WeightingMethod = "NASC": A table linking AcousticCategory with the LengthExponent used to convert from NASC to density.

Value

An object of StoX data type BioticAssignment.

Details

The BioStationWeighting function is used to update the weighting variables of the biotic stations that are associated in BioticAssignment. The list of assigned biotic hauls and weighting variables of an assignment, will in another function be used to make a total combined length frequency distribution from all the individual haul distributions.

A set of automatic WeightingMethods are available to update the haul weighing variables. Note that the weighting may change if an additional species is included for all WeightingMethods except "Equal":

Equal

All assigned biotic hauls are given equal weight by assigning the value 1 to the weighting variables.

NumberOfLengthSamples The assigned biotic hauls are given a weighting value according to the number of individual length samples of the target species at the biotic station. The parameter MaxNumberOfLengthSamples is also associated with this method and is used to limit the weighting to a maximum number of length samplesof a haul. Note that the weighting may change if an additional species is included.

NormalizedTotalWeight

The assigned biotic hauls are given a weighting value according to the normalized catch weight of the target species at the station. The weighting value is calculated as catch weight divided by towing distance. This normalization makes the stations comparable regardless of catch effort.

NormalizedTotalCount

The assigned biotic hauls are given a weighting value according to the normalized catch count (number of individuals) of the target species at the biotic station. The weighting value is calculated as catch count divided by towing distance. This normalization makes the stations comparable regardless of catch effort.

SumWeightedCount

The assigned biotic hauls are given a weighting value according to the estimated normalized length distribution count (number of individuals in all length groups) of the target species at the biotic station. It is a requirement that the lengthdistribution data is of distribution type Standard or Normalized (normalized to one nautical mile towing distance).

InverseSumWeightedCount

The assigned biotic hauls are given a weighting value as the inverse of the sum of the count of all length groups and all species. The weighting value \(w_b\) is calculated as:

$$w_b = \frac{1}{\sum_{s_b}^{n_b} \sum_{l=1}^{m_{s,b}} c_{l,s,b} }$$

where:

\(w_b\) = weighting value of biotic haul b

\(s_b\) = species in the biotic haul b

\(n_b\) = number of species in the input data off biotic haul b

\(l\) = length group number

\(m_{s,b}\) = number of length groups for species s in biotic haul b

\(c_{l,s,b}\) = count in length group l for species s in biotic haul b

The method is commonly used in split NASC (nautical area scattering coefficient) models to split an acoustic category of several species by using the length distributions of the these species. The sum of the splitted NASC values of all the species will be equal to the NASC of the original combined acoustic multispecies category. By multiplying the calculated weighting value from this method, by the original (input) numbers in each length group for all species, a relative station length distribution can later be made (sum of length groups for all species is 1) and used in the split NASC process.

It is a requirement that the lengthdistribution data is of distribution type Standard or Normalized (normalized to one nautical mile towing distance).

NASC Not yet implemented.

The assigned biotic hauls are given weighting variable values with the basis in the surrounding NASC values. By combining these NASC values with the length distribution of the biotic haul, a density as number of fish per square nautical mile is calculated and used as the weighting variable value for each biotic haul.

A search for acoustic NASC values (at EDSU resolution) is performed within a given radius around a biotic station. A weighted (by integrator distance of the EDSUs) mean NASC is calculated from the surrounding NASC values and this is used in the further weighting value calculations. Using this combined NASC value, the length distribution of the biotic station and a target strength (TS) versus length empirical relationship, the weighting variable density of the biotic station is first calculated by length group. The sum of densities (number per square nautical mile) for all length groups of the target species at the given biotic haul, is than calculated and applied as the weighting variable for the biotic haul.Note that if an EDSU NASC value is used for assignment to several biotic stations, the NASC value is split and devided between these biotic stations.

The NASC WeightingMethod is associated with the following user parameters:

Radius: Search radius (nautical miles) for NASC values (at EDSU resolution) around a biotic station

LengthExponent: LengthExponent in the target stregth versus length formula.

TargetStrength0: TargetStrength0 in the target stregth versus length formula as:

$$TS = LengthExponent \log_{10}{l} + TargetStrength0$$

where l is the fish "total length" in centimeters given as the lower value of the length group interval.

See also