Biotic PSU

This function defines the BioticPSU process data, linking strata, biotic PSUs and Stations

DefineBioticPSU(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxBioticData,
  DefinitionMethod = c("StationToPSU", "None")
)

Arguments

processData
The current data produced by a previous instance of the function.

UseProcessData
Logical: If TRUE use the existing function output in the process.

StratumPolygon
The StratumPolygon process data.

StoxBioticData
The StoxBioticData data (defined in RstoxData).

DefinitionMethod
Character: A string naming the method to use, one of "StationToPSU", which sets each Station as a PSU, and "None" for pure manual actions by the user.

Value

An object of StoX data type BioticPSU.

Details

This function is awesome and does excellent stuff.

See also

Examples

x <- 1