Mean NASC

This function averages SumNASCData horizontally to the acoustic PSU resolution. Optionally, NASCData can be used as input, in which case SumNASC is run first.

MeanNASC(
  NASCData,
  SumNASCData,
  LayerDefinition = c("FunctionParameter", "FunctionInput", "PreDefined"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "Table"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  AcousticLayer = NULL,
  SurveyDefinition = c("FunctionParameter", "FunctionInput"),
  SurveyDefinitionMethod = c("AllStrata", "Table"),
  SurveyTable = data.table::data.table(),
  Survey = NULL,
  PSUDefinition = c("FunctionParameter", "FunctionInput"),
  PSUDefinitionMethod = c("EDSUToPSU"),
  StratumPolygon = NULL,
  AcousticPSU = NULL
)

Arguments

NASCData

The NASCData data.

SumNASCData

The SumNASCData data.

LayerDefinition

The method to use for defining the Layers, one of FunctionParameter to define the Layers on the fly in this function, FunctionInput to import Layer process data from a previously run process using the function DefineAcousticLayer, or PreDefined whihc requires SumNASCData as input.

LayerDefinitionMethod

See DefineAcousticLayer

Resolution

Numeric: A single numeric giving the thickness of the layers.

LayerTable

A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.

AcousticLayer

The AcousticLayer process data.

SurveyDefinition

The method to use for defining the Survey, one of FunctionParameter to define the Survey on the fly in this function, or FunctionInput to import Survey process data from a previously run process by the input Survey.

SurveyDefinitionMethod

See DefineSurvey

SurveyTable

A table of the two columns Stratum and Survey.

Survey

The Survey process data.

PSUDefinition

The method to use for defining the PSUs, one of FunctionParameter to define the PSUs on the fly in this function, or FunctionInput to import PSU process data from a previously run process by AcousticPSU.

PSUDefinitionMethod

See DefineAcousticPSU

StratumPolygon

The StratumPolygon process data.

AcousticPSU

The AcousticPSU process data.

See also