Model data used by RstoxBase

The model data of a StoX model are the data generated during the model run based on input data and user settings and resources given in the project description (project.json file). Model data are transient and only exists from a process has been run until the project is closed.

Arguments

BioticData
The BioticData data (defined in RstoxData).

AcousticData
The AcousticData data (defined in RstoxData).

LandingData
The LandingData data (defined in RstoxData).

StoxBioticData
The StoxBioticData data (defined in RstoxData).

StoxAcousticData
The StoxAcousticData data (defined in RstoxData).

StoxLandingData
The StoxLandingData data (defined in RstoxData).

StratumAreaData
The StratumAreaData data.

NASCData
The NASCData data.

SumNASCData
The SumNASCData data.

MeanNASCData
The MeanNASCData data.

LengthDistributionData
The LengthDistributionData data.

SumLengthDistributionData
The SumLengthDistributionData data.

MeanLengthDistributionData
The MeanLengthDistributionData data.

AssignmentLengthDistributionData
The AssignmentLengthDistributionData data.

DensityData
The DensityData data.

MeanDensityData
The MeanDensityData data.

SpeciesCategoryCatchData
The SpeciesCategoryCatchData data.

AbundanceData
The AbundanceData data.

IndividualsData
The IndividualsData data.

SuperIndividualsData
The SuperIndividualsData data.

ReportSuperIndividualsData
The ReportSuperIndividualsData data.

See also