Model data used by RstoxBase

The model data of a StoX model are the data generated during the model run based on input data and user settings and resources given in the project description (project.json file). Model data are transient and only exists from a process has been run until the project is closed.

Arguments

StratumAreaData

The StratumAreaData data.

TransectDesignData

The TransectDesignData data.

StationsAlongTransectDesignData

The StationsAlongTransectDesignData data.

ReportTransectDesignData

The ReportTransectDesignData data.

WriteTransectDesignData

The WriteTransectDesignData data.

PlotTransectDesignData

The PlotTransectDesignData data.

NASCData

The NASCData data.

SumNASCData

The SumNASCData data.

MeanNASCData

The MeanNASCData data.

LengthDistributionData

The LengthDistributionData data.

SumLengthDistributionData

The SumLengthDistributionData data.

MeanLengthDistributionData

The MeanLengthDistributionData data.

AssignmentLengthDistributionData

The AssignmentLengthDistributionData data.

DensityData

The DensityData data.

ReportDensityData

The ReportDensityData data.

MeanDensityData

The MeanDensityData data.

SpeciesCategoryCatchData

The SpeciesCategoryCatchData data.

ReportSpeciesCategoryCatchData

The ReportSpeciesCategoryCatchData data.

SumSpeciesCategoryCatchData

The SumSpeciesCategoryCatchData data.

MeanSpeciesCategoryCatchData

The MeanSpeciesCategoryCatchData data.

PreySpeciesCategoryCatchData

The PreySpeciesCategoryCatchData data.

ReportPreySpeciesCategoryCatchData

The ReportPreySpeciesCategoryCatchData data.

QuantityData

The QuantityData data.

ReportQuantityData

The ReportQuantityData data.

IndividualsData

The IndividualsData data.

SuperIndividualsData

The SuperIndividualsData data.

ReportSuperIndividualsData

The ReportSuperIndividualsData data.

PlotAcousticTrawlSurveyData

The PlotAcousticTrawlSurveyData data.

BioticData

BioticData (defined in RstoxData).

AcousticData

AcousticData (defined in RstoxData).

LandingData

LandingData (defined in RstoxData).

StoxBioticData

StoxBioticData (defined in RstoxData).

StoxAcousticData

StoxAcousticData (defined in RstoxData).

StoxLandingData

StoxLandingData (defined in RstoxData).

MergeStoxBioticData

MergeStoxBioticData (defined in RstoxData).

MergeStoxAcousticData

MergeStoxAcousticData (defined in RstoxData).

ICESBioticData

ICESBioticData (defined in RstoxData).

ICESAcousticData

ICESAcousticData (defined in RstoxData).

ICESDatrasData

ICESDatrasData (defined in RstoxData).

WriteICESBioticData

WriteICESBioticData (defined in RstoxData).

WriteICESAcousticData

WriteICESAcousticData (defined in RstoxData).

WriteICESDatrasData

WriteICESDatrasData (defined in RstoxData).

See also