Plot NASCData

Plots a map with cruise line and points with size and (optionally) color representing NASC.

PlotAcousticTrawlSurvey(
  LayerDefinition = c("FunctionParameter", "FunctionInput", "PreDefined"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "Table"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  AcousticLayer = NULL,
  NASCData,
  SumNASCData,
  AcousticPSU,
  ColorVariable = character(),
  UseAcousticPSU = FALSE,
  UseDefaultAcousticPSUSettings = TRUE,
  AcousticPSULabelSize = numeric(),
  AcousticPSULabelColor = character(),
  AcousticPSULabelPosition = c("mean", "atMinLongitude", "atMaxLongitude",
    "atMinLatitude", "atMaxLatitude"),
  AcousticPSULabelHjust = numeric(),
  AcousticPSULabelVjust = numeric(),
  UseDefaultColorSettings = TRUE,
  PointColor = character(),
  TrackColor = character(),
  LandColor = character(),
  BorderColor = character(),
  OceanColor = character(),
  GridColor = character(),
  UseDefaultSizeSettings = TRUE,
  MaxPointSize = numeric(),
  MinPointSize = numeric(),
  TrackSize = numeric(),
  UseDefaultAspectSettings = TRUE,
  Zoom = numeric(),
  LongitudeMin = numeric(),
  LongitudeMax = numeric(),
  LatitudeMin = numeric(),
  LatitudeMax = numeric(),
  LongitudeCenter = numeric(),
  LatitudeCenter = numeric(),
  UseDefaultTextSettings = TRUE,
  Title = character(),
  AxisTitleSize = numeric(),
  AxisTickSize = numeric(),
  LegendTitleSize = numeric(),
  LegendTextSize = numeric(),
  UseDefaultFileSettings = TRUE,
  Format = character(),
  Width = numeric(),
  Height = numeric(),
  DotsPerInch = numeric()
)

Arguments

LayerDefinition

The method to use for defining the Layers, one of FunctionParameter to define the Layers on the fly in this function, FunctionInput to import Layer process data from a previously run process using the function DefineAcousticLayer, or PreDefined whihc requires SumNASCData as input.

LayerDefinitionMethod

See DefineAcousticLayer

Resolution

Numeric: A single numeric giving the thickness of the layers.

LayerTable

A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.

AcousticLayer

The AcousticLayer process data.

NASCData

The NASCData data.

SumNASCData

The SumNASCData data.

AcousticPSU

The AcousticPSU process data.

ColorVariable

The name of the variable determining the colors of the NASC data points, defaulted to "NASC" (both size AND color reflecting the NASC values). If the variable is a categorical variable (character or integer), the discrete colors can be set by the PointColor argument. If the variable is a continuous variable (numeric), the color scale can be set by the PointColorScale argument.

UseAcousticPSU

Logical: If TRUE plot only EDSUs tagged to acoustic PSUs.

UseDefaultAcousticPSUSettings

Logical: If TRUE (default) use the default settings for indicating acoustic PSU in the plots (text position, color and size). Setting this to FALSE will show all acoustic PSU options of the plotting function in the StoX GUI.

AcousticPSULabelSize

The text size of the plotted acoustic PSU labels.

AcousticPSULabelColor

The text color of the plotted acoustic PSU labels.

AcousticPSULabelPosition

The position of the text of the plotted acoustic PSU labels, one of "mean", "atMinLongitude", "atMaxLongitude", "atMinLatitude" or "atMaxLatitude".

AcousticPSULabelHjust

A value to displace the acoustic PSU label with, where 0 means left-justified and 1 means right-justified, and values outside of [0, 1] are valid.

AcousticPSULabelVjust

value to displace the acoustic PSU label with, where 0 means bottom-justified and 1 means top-justified, and values outside of [0, 1] are valid.

UseDefaultColorSettings

Logical: If TRUE (default) use the default color settings of the plotting function. Setting this to FALSE will show all color options of the plotting function in the StoX GUI.

PointColor

The colors to use when plotting the data points. If the ColorVariable is a discrete variable, a vector of color of the same length as the number of discrete values should be given, defaulted to the default ggplot2 color palette (see the scales package, and specifically the function hue_pal for how to generate these colors). If a continuous variable the colors scale can be given either as vector of colors comprising equally spaced colors of the color scale, or as the name of a color scale function with the first argument being the number of colors. The default is the combined.color. Other options for color scale function are "rainbow", "hcl.colors", "heat.colors", "terrain.colors", "topo.colors" or "cm.colors".

TrackColor

The color to use for the vessel track, defaulted to "black".

LandColor

The color to use for the land, defaulted to the color used for land in the StoX GUI map (yellow color, rgb(253, 254, 204, maxColorValue = 255)).

BorderColor

The color to use for the borders on land, defaulted to "grey50".

OceanColor

The color to use for the (ocean) background, defaulted to "white", as in the StoX GUI map.

GridColor

The color to use for the longitude/latitude grid lines, defaulted to the color used for the longitude/latitude grid in the StoX GUI map (blue color, rgb(223, 242, 255, maxColorValue = 255).

UseDefaultSizeSettings

Logical: If TRUE (default) use the default point size settings of the plotting function. Setting this to FALSE will show all point size options of the plotting function in the StoX GUI.

MaxPointSize

The maximum size of the points, defaulted to 10.

MinPointSize

The minimum size of the points (the size of points of data with value 0), defaulted to 0.5.

TrackSize

The size of the track points, defaulted to 1.

UseDefaultAspectSettings

Logical: If TRUE (default) use the default aspect settings of the plotting function (width, height, etc.). Setting this to FALSE will show all aspect options of the plotting function in the StoX GUI.

Zoom

The zoom of the plot, defaulted to 1, which uses the range of the data as the plot limits. Higher values zooms out and lower values zooms in.

LongitudeMin

The minimum longitude limit of the plot, overriding the value derived from Zoom.

LongitudeMax

The maximum longitude limit of the plot, overriding the value derived from Zoom.

LatitudeMin

The minimum latitude limit of the plot, overriding the value derived from Zoom.

LatitudeMax

The maximum latitude limit of the plot, overriding the value derived from Zoom.

LongitudeCenter

The longitude of the point around which the plot is zoomed usinng Zoom.

LatitudeCenter

The latitude of the point around which the plot is zoomed usinng Zoom.

UseDefaultTextSettings

Logical: If TRUE (default) use the default text settings of the plotting function. Setting this to FALSE will show all text options of the plotting function in the StoX GUI.

Title

The title of the plot, defaulted to no title.

AxisTitleSize

The size of the axis titles, defaulted to 20.

AxisTickSize

The size of the axis tick marks, defaulted to 20.

LegendTitleSize

The size of the legend title, defaulted to 20.

LegendTextSize

The size of the legend text, defaulted to 20.

UseDefaultFileSettings

Logical: If TRUE (default) use the default file settings of the plotting function. Setting this to FALSE will show all file options of the plotting function in the StoX GUI.

Format

The format of the output file from plotting functions, defaulted to "png". Other options are defined by ggsave, including "tiff", "jpeg" and "pdf".

Width

The width of the plot in centimeters, defaulted to 17, as per the ICES Journal og Marine Science author recommendations.

Height

The height of the plot in centimeters, defaulted to 17.

DotsPerInch

The number of dots per inch (DPI), defaulted to 1200.

Value

A PlotAcousticTrawlSurveyData object.