Swept-area density

This function calculates the area density of fish as number of individuals or weight (kg) per square nautical mile, as determined by the DensityType.

SweptAreaDensity(
  SweptAreaDensityMethod = c("LengthDistributed", "TotalCatch"),
  MeanLengthDistributionData,
  MeanSpeciesCategoryCatchData,
  SweepWidthMethod = c("Constant", "PreDefined"),
  SweepWidth = double(),
  DensityType = character()
)

Arguments

SweptAreaDensityMethod

The method to use for the swept-area calculation, one of \"LengthDistributed\" for calculating density from the length distribution (MeanLengthDistributionData), and \"TotalCatch\" for calculating density from the total catch (MeanSpeciesCategoryCatchData).

MeanLengthDistributionData

The MeanLengthDistributionData data.

MeanSpeciesCategoryCatchData

The MeanSpeciesCategoryCatchData data.

SweepWidthMethod

The method for calculating the sweep width. Possible options are (1) "Constant", which requires SweepWidth to be set as the constant sweep width, and (2) "PreDefined", impying that the sweep width is already incorporated in the WeightedNumber in the MeanLengthDistributionData using link{GearDependentLengthDistributionCompensation} or link{LengthDependentLengthDistributionCompensation}, or in the MeanSpeciesCategoryCatchData using link{GearDependentSpeciesCategoryCatchCompensation}.

SweepWidth

The constant sweep width in meters.

DensityType

The requested density type, currently only "AreaNumberDensity" is supported for SweptAreaDensityMethod = "LengthDistributed", and one of "AreaNumberDensity" and "AreaWeightDensity" (kg) for SweptAreaDensityMethod = "TotalCatch". All area densities are given per square nautical mile.

Value

An object of StoX data type DensityData. Note that the Density column of the Data table has unit kg nmi^-2 if SweptAreaDensityMethod is "TotalCatch" and DensityType is "AreaWeightensity". See also QuantityData

See also