Function to aggregate baseline data

Function to aggregate baseline data

aggregateBaselineDataOneTable(
  stoxData,
  TargetVariable,
  aggregationFunction = getReportFunctions(),
  subTable = character(),
  GroupingVariables = character(),
  InformationVariables = character(),
  na.rm = FALSE,
  padWithZerosOn = character(),
  WeightingVariable = character()
)

Arguments

stoxData

Output from any StoX function.

TargetVariable

The variable to report.

aggregationFunction

The function to apply, see RstoxBase::getRstoxBaseDefinitions("reportFunctions")$functionName.

subTable

The name of the sub table to aggregate on, if stoxData is a list of tables.

GroupingVariables

The variables to report by. For most applications GroupingVariables should include "Survey" or "SpeciesCategory", unless the user needs to sum over all Survey and SpeciesCategory.

InformationVariables

Variables to include as columns to the end of the report table. These cannot have more unique combinations than the GroupingVariables.

na.rm

Used in the function specified by aggregationFunction.

padWithZerosOn

Character vector giving the variables for which missing values should be padded with zeros. This is used particularly for bootstrapping, where a fish length missing in a bootstrap run should be considered as samples with zero individuals, and not missing, so that summary statistics end up taking all bootstrap replicates into account (if not a mean would be overestimated). When padWithZerosOn has positive length, padding with zeros is applied to this variable and to the GroupingVariables.

WeightingVariable

The variable to weight by. Only relevant for ReportFunction "weighted.mean".

Value

An aggregated version of the input stoxData.

See also