Get RstoxBase definitions

This function gets vital definitions from the RstoxBase environment.

getRstoxBaseDefinitions(name = NULL, ...)

Arguments

name
An optional string vector denoting which definitions to extract.

...
values overriding the values of definitions.

Value

A list of definitions.

Examples

getRstoxBaseDefinitions()
#> $allResolution
#> $allResolution$NASCData
#> [1] "NASC"
#> 
#> $allResolution$SumNASCData
#> [1] "NASC"
#> 
#> $allResolution$MeanNASCData
#> [1] "NASC"
#> 
#> $allResolution$LengthDistributionData
#> [1] "LengthDistribution"
#> 
#> $allResolution$SumLengthDistributionData
#> [1] "LengthDistribution"
#> 
#> $allResolution$MeanLengthDistributionData
#> [1] "LengthDistribution"
#> 
#> $allResolution$DensityData
#> [1] "Density"
#> 
#> $allResolution$MeanDensityData
#> [1] "Density"
#> 
#> $allResolution$AbundanceData
#> [1] "Abundance"
#> 
#> 
#> $dataTypeDefinition
#> $dataTypeDefinition$NASCData
#> $dataTypeDefinition$NASCData$horizontalResolution
#> [1] "EDSU"
#> 
#> $dataTypeDefinition$NASCData$verticalResolution
#> [1] "Channel"
#> 
#> $dataTypeDefinition$NASCData$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$NASCData$categoryVariable
#> [1] "AcousticCategory"
#> 
#> $dataTypeDefinition$NASCData$groupingVariables
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$NASCData$data
#> [1] "NASC"
#> 
#> $dataTypeDefinition$NASCData$verticalRawDimension
#> [1] "MinChannelDepth" "MaxChannelDepth"
#> 
#> $dataTypeDefinition$NASCData$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$NASCData$weighting
#> [1] "NASCWeight"
#> 
#> $dataTypeDefinition$NASCData$type
#> [1] "ChannelReferenceType"
#> 
#> $dataTypeDefinition$NASCData$other
#> [1] "ChannelReferenceDepth" "ChannelReferenceTilt"  "Cruise"               
#> [4] "EffectiveLogDistance"  "DateTime"              "Longitude"            
#> [7] "Latitude"             
#> 
#> 
#> $dataTypeDefinition$SumNASCData
#> $dataTypeDefinition$SumNASCData$Data
#> $dataTypeDefinition$SumNASCData$Data$horizontalResolution
#> [1] "EDSU"
#> 
#> $dataTypeDefinition$SumNASCData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$SumNASCData$Data$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$SumNASCData$Data$categoryVariable
#> [1] "AcousticCategory"
#> 
#> $dataTypeDefinition$SumNASCData$Data$groupingVariables
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$SumNASCData$Data$data
#> [1] "NASC"
#> 
#> $dataTypeDefinition$SumNASCData$Data$verticalRawDimension
#> NULL
#> 
#> $dataTypeDefinition$SumNASCData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$SumNASCData$Data$weighting
#> [1] "SumNASCWeight"
#> 
#> $dataTypeDefinition$SumNASCData$Data$type
#> [1] "ChannelReferenceType"
#> 
#> $dataTypeDefinition$SumNASCData$Data$other
#> [1] "ChannelReferenceDepth" "ChannelReferenceTilt"  "Cruise"               
#> [4] "EffectiveLogDistance"  "DateTime"              "Longitude"            
#> [7] "Latitude"             
#> 
#> 
#> $dataTypeDefinition$SumNASCData$Resolution
#> $dataTypeDefinition$SumNASCData$Resolution$horizontalResolution
#> [1] "EDSU"
#> 
#> $dataTypeDefinition$SumNASCData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> 
#> 
#> $dataTypeDefinition$MeanNASCData
#> $dataTypeDefinition$MeanNASCData$Data
#> $dataTypeDefinition$MeanNASCData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$MeanNASCData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$categoryVariable
#> [1] "AcousticCategory"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$groupingVariables
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$data
#> [1] "NASC"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$verticalRawDimension
#> NULL
#> 
#> $dataTypeDefinition$MeanNASCData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$weighting
#> [1] "MeanNASCWeight"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$type
#> [1] "ChannelReferenceType"
#> 
#> $dataTypeDefinition$MeanNASCData$Data$other
#> [1] "ChannelReferenceDepth" "ChannelReferenceTilt" 
#> 
#> 
#> $dataTypeDefinition$MeanNASCData$Resolution
#> $dataTypeDefinition$MeanNASCData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "EDSU"   
#> 
#> $dataTypeDefinition$MeanNASCData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> $dataTypeDefinition$MeanNASCData$Resolution$other
#> [1] "Cruise"               "EffectiveLogDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"            
#> 
#> 
#> 
#> $dataTypeDefinition$LengthDistributionData
#> $dataTypeDefinition$LengthDistributionData$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$LengthDistributionData$verticalResolution
#> [1] "Haul"
#> 
#> $dataTypeDefinition$LengthDistributionData$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$LengthDistributionData$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$LengthDistributionData$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$LengthDistributionData$data
#> [1] "WeightedCount"
#> 
#> $dataTypeDefinition$LengthDistributionData$verticalRawDimension
#> [1] "MinHaulDepth" "MaxHaulDepth"
#> 
#> $dataTypeDefinition$LengthDistributionData$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$LengthDistributionData$weighting
#> [1] "LengthDistributionWeight"
#> 
#> $dataTypeDefinition$LengthDistributionData$type
#> [1] "LengthDistributionType"
#> 
#> $dataTypeDefinition$LengthDistributionData$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"             "Gear"                
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> $dataTypeDefinition$SumLengthDistributionData
#> $dataTypeDefinition$SumLengthDistributionData$Data
#> $dataTypeDefinition$SumLengthDistributionData$Data$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$data
#> [1] "WeightedCount"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$verticalRawDimension
#> NULL
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$weighting
#> [1] "SumLengthDistributionWeight"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$type
#> [1] "LengthDistributionType"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Data$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"            
#> 
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Resolution
#> $dataTypeDefinition$SumLengthDistributionData$Resolution$horizontalResolution
#> [1] "Station"
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Resolution$verticalResolution
#> [1] "Layer" "Haul" 
#> 
#> $dataTypeDefinition$SumLengthDistributionData$Resolution$other
#> [1] "Gear"                 "VerticalNetOpening"   "HorizontalNetOpening"
#> [4] "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$MeanLengthDistributionData
#> $dataTypeDefinition$MeanLengthDistributionData$Data
#> $dataTypeDefinition$MeanLengthDistributionData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$data
#> [1] "WeightedCount"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$verticalRawDimension
#> NULL
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$weighting
#> [1] "MeanLengthDistributionWeight"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$type
#> [1] "LengthDistributionType"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Resolution
#> $dataTypeDefinition$MeanLengthDistributionData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "Station"
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Resolution$verticalResolution
#> [1] "Layer" "Haul" 
#> 
#> $dataTypeDefinition$MeanLengthDistributionData$Resolution$other
#> [1] "Cruise"               "EffectiveTowDistance" "DateTime"            
#> [4] "Longitude"            "Latitude"             "Gear"                
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData
#> $dataTypeDefinition$AssignmentLengthDistributionData$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$obserationVariable
#> NULL
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$data
#> [1] "WeightedCount"
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$weighting
#> NULL
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$type
#> [1] "LengthDistributionType"
#> 
#> $dataTypeDefinition$AssignmentLengthDistributionData$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$DensityData
#> $dataTypeDefinition$DensityData$Data
#> $dataTypeDefinition$DensityData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$DensityData$Data$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$DensityData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$DensityData$Data$obserationVariable
#> [1] "Beam"
#> 
#> $dataTypeDefinition$DensityData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$DensityData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"      "Beam"                 
#> [4] "Frequency"            
#> 
#> $dataTypeDefinition$DensityData$Data$groupingVariables_biotic
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$DensityData$Data$groupingVariables_acoustic
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$DensityData$Data$data
#> [1] "Density"
#> 
#> $dataTypeDefinition$DensityData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$DensityData$Data$weighting
#> [1] "DensityWeight"
#> 
#> $dataTypeDefinition$DensityData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$DensityData$Resolution
#> $dataTypeDefinition$DensityData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "EDSU"    "Station"
#> 
#> $dataTypeDefinition$DensityData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> $dataTypeDefinition$DensityData$Resolution$other
#> [1] "Cruise"               "EffectiveLogDistance" "EffectiveTowDistance"
#> [4] "DateTime"             "Longitude"            "Latitude"            
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$MeanDensityData
#> $dataTypeDefinition$MeanDensityData$Data
#> $dataTypeDefinition$MeanDensityData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$horizontalResolution
#> [1] "Stratum"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"      "Beam"                 
#> [4] "Frequency"            
#> 
#> $dataTypeDefinition$MeanDensityData$Data$groupingVariables_biotic
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$MeanDensityData$Data$groupingVariables_acoustic
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$data
#> [1] "Density"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$weighting
#> [1] "MeanDensityWeight"
#> 
#> $dataTypeDefinition$MeanDensityData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$MeanDensityData$Resolution
#> $dataTypeDefinition$MeanDensityData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "EDSU"    "Station"
#> 
#> $dataTypeDefinition$MeanDensityData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> $dataTypeDefinition$MeanDensityData$Resolution$other
#> [1] "Cruise"               "EffectiveLogDistance" "EffectiveTowDistance"
#> [4] "DateTime"             "Longitude"            "Latitude"            
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$AbundanceData
#> $dataTypeDefinition$AbundanceData$Data
#> $dataTypeDefinition$AbundanceData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$AbundanceData$Data$horizontalResolution
#> [1] "Stratum"
#> 
#> $dataTypeDefinition$AbundanceData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$AbundanceData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$AbundanceData$Data$groupingVariables
#> [1] "IndividualTotalLength" "LengthResolution"      "Beam"                 
#> [4] "Frequency"            
#> 
#> $dataTypeDefinition$AbundanceData$Data$groupingVariables_biotic
#> [1] "IndividualTotalLength" "LengthResolution"     
#> 
#> $dataTypeDefinition$AbundanceData$Data$groupingVariables_acoustic
#> [1] "Beam"      "Frequency"
#> 
#> $dataTypeDefinition$AbundanceData$Data$data
#> [1] "Abundance"
#> 
#> $dataTypeDefinition$AbundanceData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$AbundanceData$Data$weighting
#> NULL
#> 
#> $dataTypeDefinition$AbundanceData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$AbundanceData$Resolution
#> $dataTypeDefinition$AbundanceData$Resolution$horizontalResolution
#> [1] "Stratum" "PSU"     "EDSU"    "Station"
#> 
#> $dataTypeDefinition$AbundanceData$Resolution$verticalResolution
#> [1] "Layer"   "Channel"
#> 
#> $dataTypeDefinition$AbundanceData$Resolution$other
#> [1] "Cruise"               "EffectiveLogDistance" "EffectiveTowDistance"
#> [4] "DateTime"             "Longitude"            "Latitude"            
#> [7] "VerticalNetOpening"   "HorizontalNetOpening" "TrawlDoorSpread"     
#> 
#> 
#> 
#> $dataTypeDefinition$IndividualsData
#> $dataTypeDefinition$IndividualsData$Data
#> $dataTypeDefinition$IndividualsData$Data$horizontalResolution
#> [1] "Stratum"
#> 
#> $dataTypeDefinition$IndividualsData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$IndividualsData$Data$data
#> NULL
#> 
#> $dataTypeDefinition$IndividualsData$Data$verticalLayerDimension
#> NULL
#> 
#> $dataTypeDefinition$IndividualsData$Data$weighting
#> NULL
#> 
#> $dataTypeDefinition$IndividualsData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$IndividualsData$VariableNames
#> list()
#> 
#> 
#> $dataTypeDefinition$SuperIndividualsData
#> $dataTypeDefinition$SuperIndividualsData$Data
#> $dataTypeDefinition$SuperIndividualsData$Data$surveyDefinition
#> [1] "Survey"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$horizontalResolution
#> [1] "Stratum"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$categoryVariable
#> [1] "SpeciesCategory"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$groupingVariables
#>  [1] "CruiseKey"             "StationKey"            "HaulKey"              
#>  [4] "SpeciesCategoryKey"    "SampleKey"             "IndividualKey"        
#>  [7] "IndividualTotalLength" "LengthResolution"      "IndividualRoundWeight"
#> [10] "Beam"                  "Frequency"            
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$data
#> [1] "Abundance" "Biomass"  
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$verticalLayerDimension
#> [1] "MinLayerDepth" "MaxLayerDepth"
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$weighting
#> NULL
#> 
#> $dataTypeDefinition$SuperIndividualsData$Data$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$SuperIndividualsData$VariableNames
#> list()
#> 
#> 
#> $dataTypeDefinition$BioticAssignment
#> $dataTypeDefinition$BioticAssignment$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$BioticAssignment$verticalResolution
#> [1] "Layer"
#> 
#> $dataTypeDefinition$BioticAssignment$categoryVariable
#> [1] "Haul"
#> 
#> $dataTypeDefinition$BioticAssignment$weighting
#> [1] "WeightingFactor"
#> 
#> $dataTypeDefinition$BioticAssignment$other
#> NULL
#> 
#> 
#> $dataTypeDefinition$BioticPSU
#> $dataTypeDefinition$BioticPSU$Stratum_PSU
#> $dataTypeDefinition$BioticPSU$Stratum_PSU$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> 
#> $dataTypeDefinition$BioticPSU$Station_PSU
#> $dataTypeDefinition$BioticPSU$Station_PSU$horizontalResolution
#> [1] "Station" "PSU"    
#> 
#> 
#> 
#> $dataTypeDefinition$AcousticPSU
#> $dataTypeDefinition$AcousticPSU$Stratum_PSU
#> $dataTypeDefinition$AcousticPSU$Stratum_PSU$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> 
#> $dataTypeDefinition$AcousticPSU$EDSU_PSU
#> $dataTypeDefinition$AcousticPSU$EDSU_PSU$horizontalResolution
#> [1] "EDSU" "PSU" 
#> 
#> 
#> $dataTypeDefinition$AcousticPSU$PSUByTime
#> $dataTypeDefinition$AcousticPSU$PSUByTime$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $dataTypeDefinition$AcousticPSU$PSUByTime$categoryVariable
#> [1] "Cruise"
#> 
#> $dataTypeDefinition$AcousticPSU$PSUByTime$groupingVariables
#> [1] "StartDateTime" "StopDateTime" 
#> 
#> 
#> 
#> 
#> $dataTypeUnits
#> $dataTypeUnits$NASCData
#> $dataTypeUnits$NASCData$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$NASCData$NASC
#> [1] "sA"
#> 
#> $dataTypeUnits$NASCData$MinChannelDepth
#> [1] "m"
#> 
#> $dataTypeUnits$NASCData$MaxChannelDepth
#> [1] "m"
#> 
#> $dataTypeUnits$NASCData$ChannelReferenceDepth
#> [1] "m"
#> 
#> $dataTypeUnits$NASCData$ChannelReferenceTilt
#> [1] "degrees"
#> 
#> $dataTypeUnits$NASCData$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$NASCData$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$NASCData$Latitude
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$SumNASCData
#> $dataTypeUnits$SumNASCData$Data
#> $dataTypeUnits$SumNASCData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$SumNASCData$Data$NASC
#> [1] "sA"
#> 
#> $dataTypeUnits$SumNASCData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumNASCData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumNASCData$Data$ChannelReferenceDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumNASCData$Data$ChannelReferenceTilt
#> [1] "degrees"
#> 
#> $dataTypeUnits$SumNASCData$Data$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$SumNASCData$Data$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$SumNASCData$Data$Latitude
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$SumNASCData$Resolution
#> list()
#> 
#> 
#> $dataTypeUnits$MeanNASCData
#> $dataTypeUnits$MeanNASCData$Data
#> $dataTypeUnits$MeanNASCData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$MeanNASCData$Data$NASC
#> [1] "sA"
#> 
#> $dataTypeUnits$MeanNASCData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanNASCData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanNASCData$Data$ChannelReferenceDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanNASCData$Data$ChannelReferenceTilt
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$MeanNASCData$Resolution
#> $dataTypeUnits$MeanNASCData$Resolution$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanNASCData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanNASCData$Resolution$Latitude
#> [1] "degrees"
#> 
#> 
#> 
#> $dataTypeUnits$LengthDistributionData
#> $dataTypeUnits$LengthDistributionData$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$LengthDistributionData$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$LengthDistributionData$WeightedCount
#> [1] "individual"
#> 
#> $dataTypeUnits$LengthDistributionData$MinHaulDepth
#> [1] "m"
#> 
#> $dataTypeUnits$LengthDistributionData$MaxHaulDepth
#> [1] "m"
#> 
#> $dataTypeUnits$LengthDistributionData$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$LengthDistributionData$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$LengthDistributionData$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$LengthDistributionData$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$LengthDistributionData$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$LengthDistributionData$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$SumLengthDistributionData
#> $dataTypeUnits$SumLengthDistributionData$Data
#> $dataTypeUnits$SumLengthDistributionData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$WeightedCount
#> [1] "individual"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Data$Latitude
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$SumLengthDistributionData$Resolution
#> $dataTypeUnits$SumLengthDistributionData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$SumLengthDistributionData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$MeanLengthDistributionData
#> $dataTypeUnits$MeanLengthDistributionData$Data
#> $dataTypeUnits$MeanLengthDistributionData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$WeightedCount
#> [1] "individual"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Data$Latitude
#> [1] "degrees"
#> 
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$MeanLengthDistributionData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$AssignmentLengthDistributionData
#> $dataTypeUnits$AssignmentLengthDistributionData$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$AssignmentLengthDistributionData$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$AssignmentLengthDistributionData$WeightedCount
#> [1] "individual"
#> 
#> 
#> $dataTypeUnits$DensityData
#> $dataTypeUnits$DensityData$Data
#> $dataTypeUnits$DensityData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$DensityData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$DensityData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$DensityData$Data$Density
#> [1] "nautical_areal_number_density"
#> 
#> $dataTypeUnits$DensityData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$DensityData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$DensityData$Resolution
#> $dataTypeUnits$DensityData$Resolution$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$DensityData$Resolution$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$DensityData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$DensityData$Resolution$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$DensityData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$DensityData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$DensityData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$MeanDensityData
#> $dataTypeUnits$MeanDensityData$Data
#> $dataTypeUnits$MeanDensityData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$MeanDensityData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$MeanDensityData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$MeanDensityData$Data$Density
#> [1] "nautical_areal_number_density"
#> 
#> $dataTypeUnits$MeanDensityData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$MeanDensityData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$MeanDensityData$Resolution
#> $dataTypeUnits$MeanDensityData$Resolution$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$MeanDensityData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$AbundanceData
#> $dataTypeUnits$AbundanceData$Data
#> $dataTypeUnits$AbundanceData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$AbundanceData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$AbundanceData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$AbundanceData$Data$Abundance
#> [1] "ind"
#> 
#> $dataTypeUnits$AbundanceData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$AbundanceData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$AbundanceData$Resolution
#> $dataTypeUnits$AbundanceData$Resolution$EffectiveLogDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$AbundanceData$Resolution$EffectiveTowDistance
#> [1] "nautical_mile"
#> 
#> $dataTypeUnits$AbundanceData$Resolution$Longitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$AbundanceData$Resolution$Latitude
#> [1] "degrees"
#> 
#> $dataTypeUnits$AbundanceData$Resolution$VerticalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$AbundanceData$Resolution$HorizontalNetOpening
#> [1] "m"
#> 
#> $dataTypeUnits$AbundanceData$Resolution$TrawlDoorSpread
#> [1] "m"
#> 
#> 
#> 
#> $dataTypeUnits$IndividualsData
#> list()
#> 
#> $dataTypeUnits$SuperIndividualsData
#> $dataTypeUnits$SuperIndividualsData$Data
#> $dataTypeUnits$SuperIndividualsData$Data$IndividualTotalLength
#> [1] "cm"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$LengthResolution
#> [1] "cm"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$IndividualRoundWeight
#> [1] "g"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$Frequency
#> [1] "Hz"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$Abundance
#> [1] "ind"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$Biomass
#> [1] "g"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$MinLayerDepth
#> [1] "m"
#> 
#> $dataTypeUnits$SuperIndividualsData$Data$MaxLayerDepth
#> [1] "m"
#> 
#> 
#> $dataTypeUnits$SuperIndividualsData$VariableNames
#> list()
#> 
#> 
#> $dataTypeUnits$BioticAssignment
#> list()
#> 
#> $dataTypeUnits$BioticPSU
#> list()
#> 
#> $dataTypeUnits$AcousticPSU
#> list()
#> 
#> 
#> $emptyStratumPolygon
#> An object of class "SpatialPolygonsDataFrame"
#> Slot "data":
#> data frame with 0 columns and 0 rows
#> 
#> Slot "polygons":
#> list()
#> 
#> Slot "plotOrder":
#> integer(0)
#> 
#> Slot "bbox":
#>              min           max
#> x -1.797693e+308 1.797693e+308
#> y -1.797693e+308 1.797693e+308
#> 
#> Slot "proj4string":
#> CRS arguments: NA 
#> 
#> 
#> $getPSUPrefix
#> function (PSUType) 
#> {
#>     if (PSUType == "Acoustic") {
#>         prefix <- "PSU"
#>     }
#>     else if (PSUType == "Biotic") {
#>         prefix <- "PSU"
#>     }
#>     else {
#>         stop("PSUType must be one of \"Acoustic\" or \"Biotic\"")
#>     }
#>     return(prefix)
#> }
#> <bytecode: 0x7fdbbe8f0c48>
#> <environment: 0x7fdbbe8ed850>
#> 
#> $getSSULabel
#> function (PSUType) 
#> {
#>     if (PSUType == "Acoustic") {
#>         SSULabel <- "EDSU"
#>     }
#>     else if (PSUType == "Biotic") {
#>         SSULabel <- "Station"
#>     }
#>     else {
#>         stop("PSUType must be one of \"Acoustic\" or \"Biotic\"")
#>     }
#>     return(SSULabel)
#> }
#> <bytecode: 0x7fdbbe8ef770>
#> <environment: 0x7fdbbe8ed850>
#> 
#> $getStationLevel
#> function (PSUType) 
#> {
#>     if (PSUType == "Acoustic") {
#>         SSULabel <- "Log"
#>     }
#>     else if (PSUType == "Biotic") {
#>         SSULabel <- "Station"
#>     }
#>     else {
#>         stop("PSUType must be one of \"Acoustic\" or \"Biotic\"")
#>     }
#>     return(SSULabel)
#> }
#> <bytecode: 0x7fdbbe8ec388>
#> <environment: 0x7fdbbe8ed850>
#> 
#> $nauticalMileInMeters
#> [1] 1852
#> 
#> $proj4string
#> [1] "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
#> 
#> $PSUByTime
#> $PSUByTime$horizontalResolution
#> [1] "Stratum" "PSU"    
#> 
#> $PSUByTime$categoryVariable
#> [1] "Cruise"
#> 
#> $PSUByTime$groupingVariables
#> [1] "StartDateTime" "StopDateTime" 
#> 
#> 
#> $RemoveMissingValuesWarning
#> [1] "StoX: Using RemoveMissingValues = TRUE implies the risk of under-estimation. E.g., if RemoveMissingValues = TRUE and a super-individual lacks IndividualRoundWeight, Biomass will be NA, and the portion of Abundance distributed to that super-individual will be excluded when summing Biomass (but included when summing Abundance). It is advised to always run with RemoveMissingValues = FALSE first, and make a thorough investigation to identify the source of any missing values. The function ImputeSuperIndividuals can be used to impute the missing information from other super-individuals."
#> 
#> $reportFunctions
#>      functionName packageName weighted weightingParameter multiple
#>  1:   summaryStox   RstoxBase    FALSE                        TRUE
#>  2:           sum        base    FALSE                       FALSE
#>  3:          mean        base    FALSE                       FALSE
#>  4: weighted.mean       stats     TRUE                  w    FALSE
#>  5:        median       stats    FALSE                       FALSE
#>  6:           min        base    FALSE                       FALSE
#>  7:           max        base    FALSE                       FALSE
#>  8:            sd       stats    FALSE                       FALSE
#>  9:           var       stats    FALSE                       FALSE
#> 10:            cv   RstoxBase    FALSE                       FALSE
#> 
#> $resolutionClasses
#> $resolutionClasses$NASC
#> $resolutionClasses$NASC$vertical
#> [1] "Layer"   "Channel"
#> 
#> $resolutionClasses$NASC$horizontal
#> [1] "Stratum" "PSU"     "EDSU"   
#> 
#> 
#> $resolutionClasses$LengthDistribution
#> $resolutionClasses$LengthDistribution$vertical
#> [1] "Layer" "Haul" 
#> 
#> $resolutionClasses$LengthDistribution$horizontal
#> [1] "Stratum" "PSU"     "Station"
#> 
#> 
#> $resolutionClasses$Density
#> $resolutionClasses$Density$vertical
#> [1] "Layer"
#> 
#> $resolutionClasses$Density$horizontal
#> [1] "Stratum" "PSU"    
#> 
#> 
#> $resolutionClasses$Abundance
#> $resolutionClasses$Abundance$vertical
#> [1] "Layer"
#> 
#> $resolutionClasses$Abundance$horizontal
#> [1] "Stratum"
#> 
#> 
#> 
#> $seedSequenceLength
#> [1] 1e+07
#> 
#> $TargetStrengthFunction_LengthAndDepthDependent
#> function (midIndividualTotalLength, TargetStrength0, LengthExponent, 
#>     DepthExponent, Depth) 
#> {
#>     TargetStrength0 + LengthExponent * log10(midIndividualTotalLength) + 
#>         DepthExponent * log10(1 + Depth/10)
#> }
#> <bytecode: 0x7fdbbe8f0318>
#> <environment: 0x7fdbbe8ed850>
#> 
#> $TargetStrengthFunction_LengthDependent
#> function (midIndividualTotalLength, TargetStrength0, LengthExponent, 
#>     Depth) 
#> {
#>     TargetStrength0 + LengthExponent * log10(midIndividualTotalLength)
#> }
#> <bytecode: 0x7fdbbe8f3e70>
#> <environment: 0x7fdbbe8ed850>
#> 
#> $TargetStrengthFunction_LengthExponent
#> function (midIndividualTotalLength, LengthExponent) 
#> {
#>     LengthExponent * log10(midIndividualTotalLength)
#> }
#> <bytecode: 0x7fdbbe8f09a8>
#> <environment: 0x7fdbbe8ed850>
#> 
#> $targetStrengthMethodTypes
#> $targetStrengthMethodTypes$LengthDependent
#> [1] "Function"
#> 
#> $targetStrengthMethodTypes$LengthAndDepthDependent
#> [1] "Function"
#> 
#> $targetStrengthMethodTypes$LengthExponent
#> [1] "Function"
#> 
#> $targetStrengthMethodTypes$TargetStrengthByLength
#> [1] "Table"
#> 
#> 
#> $targetStrengthParameters
#> $targetStrengthParameters$LengthDependent
#> [1] "TargetStrength0" "LengthExponent" 
#> 
#> $targetStrengthParameters$LengthAndDepthDependent
#> [1] "TargetStrength0" "LengthExponent"  "DepthExponent"  
#> 
#> $targetStrengthParameters$LengthExponent
#> [1] "LengthExponent"
#> 
#> $targetStrengthParameters$TargetStrengthByLength
#> [1] "TargetStrength" "TotalLength"   
#> 
#>